% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{accu_samp_threshold}
\alias{accu_samp_threshold}
\title{Compute the number of sequence to obtain a given proportion of ASV in
accumulation curves}
\usage{
accu_samp_threshold(res_accuplot, threshold = 0.95)
}
\arguments{
\item{res_accuplot}{the result of the function accu_plot()}

\item{threshold}{the proportion of ASV to obtain in each samples}
}
\value{
a value for each samples of the number of sequences needed
to obtain \code{threshold} proportion of the ASV
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data("GlobalPatterns", package = "phyloseq")
GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
p <- accu_plot(GP, "SampleType", add_nb_seq = TRUE, by.fact = TRUE, step = 10)

val_threshold <- accu_samp_threshold(p)

summary(val_threshold)

# Plot the number of sequences needed to accumulate 0.95\% of ASV in 50\%, 75\%
# and 100\% of samples
p + geom_vline(xintercept = quantile(val_threshold, probs = c(0.50, 0.75, 1)))
}
\seealso{
\code{\link[=accu_plot]{accu_plot()}}
}
\author{
Adrien Taudière
}
