% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_test.R
\name{adonis_pq}
\alias{adonis_pq}
\title{Permanova on a phyloseq object}
\usage{
adonis_pq(physeq, formula, merge_sample_by = NULL)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{formula}{(required) todo}

\item{merge_sample_by}{a vector to determine
which samples to merge using the \code{\link[speedyseq:merge_samples2]{speedyseq::merge_samples2()}}
function. Need to be in \code{physeq@sam_data}}
}
\value{
The function returns an anova.cca result object with a
new column for partial R^2.
See ?adonis2 for more information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A wrapper for the \code{\link[vegan:adonis]{vegan::adonis2()}} function in the case of \code{physeq} object.
}
\examples{
data(enterotype)
clean_enterotype <- subset_samples(
  enterotype,
  !is.na(enterotype@sam_data$Enterotype)
)
adonis_pq(clean_enterotype, "SeqTech*Enterotype")
}
\author{
Adrien Taudière
}
