% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_test.R
\name{aldex_pq}
\alias{aldex_pq}
\title{Run Aldex on a phyloseq object}
\usage{
aldex_pq(physeq, bifactor = NULL, modalities = NULL, gamma = 0.5, ...)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{bifactor}{(required) The name of a column present in the \verb{@sam_data} slot
of the physeq object. Must be a character vector or a factor.}

\item{modalities}{(default NULL) A vector of modalities to keep in the analysis.
If NULL, all modalities present in bifactor are kept. Note that only two
modalities are allowed.
@param gamma (default 0.5) The value of the Dirichlet Monte-Carlo
sampling parameter.
@param ... Additional arguments passed on to \code{ALDEx2::aldex()}}
}
\value{
The result of \code{ALDEx2::aldex()}
}
\description{
Run Aldex on a phyloseq object
}
\details{
It is a wrapper of the \code{ALDEx2::aldex()} function with default gamma=0.5.\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}
}
\examples{
res_aldex <- aldex_pq(data_fungi_mini,
  bifactor = "Height",
  modalities = c("Low", "High")
)
ALDEx2::aldex.plot(res_aldex, type = "volcano")
}
\author{
Adrien Taudière
}
