% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{as_binary_otu_table}
\alias{as_binary_otu_table}
\title{Transform the otu_table of a \code{\link{phyloseq-class}} object into a
  \code{\link{phyloseq-class}} object with à binary otu_table.}
\usage{
as_binary_otu_table(physeq, min_number = 1)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{minNumber}{(default = 1): the minimum number of sequences to put
a 1 in the otu table.}
}
\value{
A \code{physeq} object with only 0/1 in the OTU table
}
\description{
Transform the otu_table of a \code{\link{phyloseq-class}} object into a
  \code{\link{phyloseq-class}} object with à binary otu_table.
}
\note{
Usefull to test if the results are not biaised by sequences bias
  that appended during PCR or NGS pipeline.
}
\examples{
data(enterotype)
enterotype_bin <- as_binary_otu_table(enterotype)
}
\author{
Adrien Taudière
}
