% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{blast_pq}
\alias{blast_pq}
\title{Blast all sequence of \code{refseq} slot of a \code{\link{phyloseq-class}}
object against a custom database.}
\usage{
blast_pq(
  physeq,
  database = NULL,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 50,
  min_cover_cut = 50,
  unique_per_seq = FALSE,
  score_filter = TRUE,
  list_no_output_query = FALSE
)
}
\arguments{
\item{database}{(required) path to a fasta file to make the blast database}

\item{blastpath}{path to blast program}

\item{id_cut}{(default: 90) cut of in identity percent to keep result}

\item{bit_score_cut}{(default: 50) cut of in bit score to keep result}

\item{min_cover_cut}{(default: 50) cut of in query cover (\%) to keep result}

\item{unique_per_seq}{(logical) if TRUE only return the first match for
each sequence in seq2search}

\item{score_filter}{(logical) does results are filter by score? If
FALSE, \code{id_cut},\code{bit_score_cut} and \code{min_cover_cut} are ignored}
}
\value{
a blast table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=blast_to_phyloseq]{blast_to_phyloseq()}} to use \code{refseq} slot as a database
}
