% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{blast_to_phyloseq}
\alias{blast_to_phyloseq}
\title{Blast some sequence against `refseq` slot of a \code{\link{phyloseq-class}}
  object.}
\usage{
blast_to_phyloseq(
  physeq,
  seq2search,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 1e-10,
  unique_per_seq = FALSE,
  score_filter = TRUE,
  list_no_output_query = FALSE
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{seq2search}{(required): path to fasta file}

\item{blastpath}{(default = NULL): path to blast program}

\item{id_cut}{(default = 90): cut of in identity percent to keep result}

\item{bit_score_cut}{(default = 1e-10): cut of in bit score to keep result}

\item{unique_per_seq}{(default = FALSE) : only return the first match for
each sequence in seq2search}

\item{score_filter}{(default = TRUE): does results are filter by score. If
FALSE, `id_cut` and `bit_score_cut` are ignored}

\item{list_no_output_query}{(default to FALSE): does the result table include
query sequences for which `blastn` does not find any correspondence}
}
\value{

}
\description{
`r lifecycle::badge("maturing")`
}
\examples{
\dontrun{
data(data_fungi)
blastpath <- "...YOUR_PATH_TO_BLAST..."
blast_to_phyloseq(data_fungi,
  seq2search = "inst/extdata/ex.fasta",
  blastpath = blastpath
)
}
}
