% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pq.R
\name{css_pq}
\alias{css_pq}
\title{Cumulative Sum Scaling (CSS) normalization of a phyloseq object}
\usage{
css_pq(physeq, log = TRUE)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{log}{(logical, default \code{TRUE}) whether to return
\eqn{\log_2(x + 1)} transformed counts (as recommended by the
metagenomeSeq authors).}
}
\value{
A new \code{\link[phyloseq]{phyloseq-class}} object with a CSS
normalised \code{otu_table}.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Wrapper around \code{metagenomeSeq::cumNorm()} / \code{metagenomeSeq::MRcounts()}
implementing Cumulative Sum Scaling (Paulson et al. 2013,
\doi{10.1038/nmeth.2658}).
}
\examples{
\dontshow{if (rlang::is_installed("metagenomeSeq")) withAutoprint(\{ # examplesIf}
data_f_css <- css_pq(data_fungi_mini)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{metagenomeSeq::cumNorm()}
}
\author{
Adrien Taudière
}
