% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fungi_mini}
\alias{data_fungi_mini}
\title{Fungal OTU in phyloseq format}
\format{
A physeq object containing 45 taxa with references sequences
described by 14 taxonomic ranks and 137 samples described by 7 sample variables:
\itemize{
\item \emph{X}: the name of the fastq-file
\item \emph{Sample_names}: the names of ... the samples
\item \emph{Treename}: the name of an tree
\item \emph{Sample_id}: identifier for each sample
\item \emph{Height}: height of the sample in the tree
\item \emph{Diameter}: diameter of the trunk
\item \emph{Time}: time since the dead of the tree
}

A physeq object containing 45 taxa with references sequences
described by 14 taxonomic ranks and 137 samples described by 7 sample variables:
\itemize{
\item \emph{X}: the name of the fastq-file
\item \emph{Sample_names}: the names of ... the samples
\item \emph{Treename}: the name of an tree
\item \emph{Sample_id}: identifier for each sample
\item \emph{Height}: height of the sample in the tree
\item \emph{Diameter}: diameter of the trunk
\item \emph{Time}: time since the dead of the tree
}
}
\usage{
data(data_fungi_mini)

data(data_fungi_mini)
}
\description{
It is a subset of the data_fungi dataset including only Basidiomycota
with more than 5000 sequences.
}
\details{
Obtain using \code{data_fungi_mini <- subset_taxa(data_fungi, Phylum == "Basidiomycota")}
and then \code{data_fungi_mini <-   subset_taxa_pq(data_fungi_mini, colSums(data_fungi_mini@otu_table) > 5000)}
}
\keyword{datasets}
