% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fungi_sp_known}
\alias{data_fungi_sp_known}
\title{Fungal OTU in phyloseq format}
\format{
A physeq object containing 651 taxa with references sequences
described by 14 taxonomic ranks and 185 samples described by 7 sample variables:
\itemize{
\item \emph{X}: the name of the fastq-file
\item \emph{Sample_names}: the names of ... the samples
\item \emph{Treename}: the name of an tree
\item \emph{Sample_id}: identifier for each sample
\item \emph{Height}: height of the sample in the tree
\item \emph{Diameter}: diameter of the trunk
\item \emph{Time}: time since the dead of the tree
}
}
\usage{
data_fungi_sp_known
}
\description{
It is a subset of the data_fungi dataset including only taxa with information
at the species level
}
\details{
Obtain using \code{data_fungi_sp_known <- subset_taxa(data_fungi, !is.na(data_fungi@tax_table[,"Species"]))}
}
\keyword{datasets}
