% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{fac2col}
\alias{fac2col}
\title{Translates a factor into colors.}
\usage{
fac2col(x, col.pal = funky_color, na.col = "grey", seed = NULL)
}
\arguments{
\item{x}{a numeric vector (for num2col) or a vector converted to a factor (for fac2col).}

\item{col.pal}{(default funky_color)    a function generating colors according to a given palette.}

\item{na.col}{(default grey) the color to be used for missing values (NAs)}

\item{seed}{(default NULL) a seed for R's random number generated, used to fix the random permutation of colors in the palette used; if NULL, no randomization is used and the colors are taken from the palette according to the ordering of the levels}
}
\value{
a color vector
}
\description{
Translates a factor into colors.
}
\seealso{
The R package RColorBrewer, proposing a nice selection of color palettes. The viridis package, with many excellent palettes
}
\author{
Thibaut Jombart in \code{adegenet} package
}
