% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmseqs2.R
\name{find_mmseqs2}
\alias{find_mmseqs2}
\title{Find the MMseqs2 binary}
\usage{
find_mmseqs2()
}
\value{
A character string with the path to the \code{mmseqs} binary.
}
\description{
Looks for the MMseqs2 binary in three places, in order:
\enumerate{
\item The option \code{MiscMetabar.mmseqs2path} (if set).
\item A local copy installed by \code{\link[=install_mmseqs2]{install_mmseqs2()}} in the user data directory.
\item The system \code{PATH}.
}
}
\seealso{
\code{\link[=install_mmseqs2]{install_mmseqs2()}}, \code{\link[=is_mmseqs2_installed]{is_mmseqs2_installed()}}, \code{\link[=assign_mmseqs2]{assign_mmseqs2()}}
}
\author{
Adrien Taudière
}
