% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{format2dada2_species}
\alias{format2dada2_species}
\title{Format a fasta database in dada2 format for Species assignment}
\usage{
format2dada2_species(
  fasta_db = NULL,
  taxnames = NULL,
  from_sintax = FALSE,
  output_path = NULL,
  ...
)
}
\arguments{
\item{fasta_db}{A link to a fasta files}

\item{taxnames}{A list of names to format. You must specify either fasta_db OR taxnames, not both.}

\item{from_sintax}{(logical, default FALSE) Is the original fasta file in sintax format?}

\item{output_path}{(optional) A path to an output fasta files. Only used if fasta_db is set.}

\item{...}{Additional arguments passed on to \code{\link[=format2sintax]{format2sintax()}} function}
}
\value{
Either an object of class DNAStringSet or a vector of reformated names
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


First format in sintax format and then in dada2 format
}
\seealso{
\code{\link[=format2dada2_species]{format2dada2_species()}}, \code{\link[=format2sintax]{format2sintax()}}
}
\author{
Adrien Taudière
}
