% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{funky_color}
\alias{funky_color}
\title{Funky palette color}
\usage{
funky_color(n)
}
\arguments{
\item{n}{a number of colors}
}
\value{
a color palette
}
\description{
Funky palette color
}
\examples{
funky_color(5)
}
\seealso{
The R package RColorBrewer, proposing a nice selection of color palettes. The viridis package, with many excellent palettes
}
\author{
Thibaut Jombart in \code{adegenet} package
}
