% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggVenn_phyloseq}
\alias{ggVenn_phyloseq}
\title{Venn diagram of \code{\link{phyloseq-class}} object using
\code{ggVennDiagram::ggVennDiagram} function}
\usage{
ggVenn_phyloseq(physeq = NULL, fact = NULL, min_nb_seq = 0)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(required): Name of the factor to cluster samples by modalities.
Need to be in \code{physeq@sam_data}.}

\item{min_nb_seq}{(Default: 0)): minimum number of sequences by OTUs by
samples to take into count this OTUs in this sample. For example,
if min_nb_seq=2,each value of 2 or less in the OTU table
will not count in the venn diagramm}
}
\value{
A \code{\link{ggplot}}2 plot representing Venn diagramm of
modalities of the argument \code{factor}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\author{
Adrien Taudière
}
