% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hill_phyloseq}
\alias{hill_phyloseq}
\title{Graphical representation of hill number 0, 1 and 2 accross a factor}
\usage{
hill_phyloseq(physeq, variable, color_fac)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}} object}

\item{variable}{(required): The variable to test}

\item{color_fac}{(optional): The variable to color the barplot}
}
\value{
A list of 4 ggplot2 plot.
  * plot_Hill_0 : the boxplot of Hill number 0 (= species richness)
    against the variable
  * plot_Hill_1 : the boxplot of Hill number 1 (= Shannon index)
     against the variable
  * plot_Hill_2 : the boxplot of Hill number 2 (= Simpson index)
    against the variable
  * plot_tuckey : plot the result of the Tuckey HSD test
}
\description{
Graphical representation of hill number 0, 1 and 2 accross a factor
}
