% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuckeyTest.R
\name{hill_tuckey_phyloseq}
\alias{hill_tuckey_phyloseq}
\title{Calculate hill number and compute Tuckey post-hoc test}
\usage{
hill_tuckey_phyloseq(physeq, modality)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{modality}{(required): the variable to test}
}
\value{
A ggplot2 object
}
\description{
Calculate hill number and compute Tuckey post-hoc test
}
\author{
Adrien Taudière
data("GlobalPatterns")
GlobalPatterns@sam_data[,'Soil_logical'] <- ifelse(GlobalPatterns@sam_data[,"SampleType"] == "Soil", "Soil", "Not Soil")
hill_tuckey_phyloseq(GlobalPatterns, "Soil_logical")
}
