% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div_test.R
\name{hill_tuckey_pq}
\alias{hill_tuckey_pq}
\title{Calculate hill number and compute Tuckey post-hoc test}
\usage{
hill_tuckey_pq(
  physeq,
  modality,
  q = c(0, 1, 2),
  hill_scales = lifecycle::deprecated(),
  silent = TRUE,
  correction_for_sample_size = TRUE,
  ...
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{modality}{(required) the variable to test}

\item{q}{(numeric vector) Hill diversity orders to compute (q values).
Default computes Hill number 0 (species richness), Hill number 1
(exponential of Shannon index) and Hill number 2 (inverse of Simpson
index). Formerly \code{hill_scales}. Hill numbers are more appropriate in DNA
metabarcoding studies when \code{q > 0} (Alberdi & Gilbert, 2019;
Calderón-Sanou et al., 2019).}

\item{hill_scales}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{q} instead.}

\item{silent}{(logical) If TRUE, no message are printing.}

\item{correction_for_sample_size}{(logical, default TRUE) This function
use a sqrt of the read numbers in the linear model in order to
correct for uneven sampling depth.}

\item{...}{Additional arguments passed to \code{\link[=divent_hill_matrix_pq]{divent_hill_matrix_pq()}} and
hence to \code{\link[divent:div_hill]{divent::div_hill()}} (e.g. \code{estimator = "naive"} to match
vegan-style results).}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}


Note that, by default, this function use a sqrt of the read numbers in the linear
model in order to correct for uneven sampling depth.
}
\examples{
data("GlobalPatterns", package = "phyloseq")
GlobalPatterns@sam_data[, "Soil_logical"] <-
  ifelse(GlobalPatterns@sam_data[, "SampleType"] == "Soil", "Soil", "Not Soil")
hill_tuckey_pq(GlobalPatterns, "Soil_logical")
hill_tuckey_pq(GlobalPatterns, "Soil_logical", q = 1:2)
}
\references{
Alberdi, A., & Gilbert, M. T. P. (2019). A guide to the application of
Hill numbers to DNA-based diversity analyses. \emph{Molecular Ecology Resources}.
\doi{10.1111/1755-0998.13014}

Calderón-Sanou, I., Münkemüller, T., Boyer, F., Zinger, L., & Thuiller, W.
(2019). From environmental DNA sequences to ecological conclusions: How
strong is the influence of methodological choices? \emph{Journal of Biogeography},
47. \doi{10.1111/jbi.13681}
}
\author{
Adrien Taudière
}
