% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{install_pkg_needed}
\alias{install_pkg_needed}
\title{Install a package if not present}
\usage{
install_pkg_needed(
  pkg,
  use_pak = TRUE,
  bioconductor_pkg = FALSE,
  github_pkg = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pkg}{The name of the package}

\item{use_pak}{(logical, default TRUE) Use of \code{pak::pkg_install()}. If FALSE
use the base \code{install.package()} function or the function
\code{BiocManager::install()} if bioconductor_pkg is true or the function}

\item{bioconductor_pkg}{(logical, default FALSE). If use_pak is TRUE,
do nothing, else use \code{BiocManager::install()} to install the package.}

\item{github_pkg}{(logical, default FALSE). If use_pak is TRUE,
do nothing, else use \code{devtools::install_github} to install the package.}

\item{verbose}{(logical, default FALSE) Does the function print message?}
}
\value{
Nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Adrien Taudière
}
