% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_blastn_installed}
\alias{is_blastn_installed}
\title{Test if blastn is installed.}
\usage{
is_blastn_installed(path = "blastn")
}
\arguments{
\item{path}{(default: blastn) Path to blastn (NCBI BLAST+)}
}
\value{
A logical that say if blastn is installed.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stable"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_blastn_installed()
}
\author{
Adrien Taudière
}
