% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_falco_installed}
\alias{is_falco_installed}
\title{Test if falco is installed.}
\usage{
is_falco_installed(path = "falco")
}
\arguments{
\item{path}{(default: falco) Path to falco}
}
\value{
A logical that say if falco is install in
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_falco_installed()
}
\author{
Adrien Taudière
}
