% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastp.R
\name{is_fastp_installed}
\alias{is_fastp_installed}
\title{Test if fastp is installed.}
\usage{
is_fastp_installed(path = "fastp")
}
\arguments{
\item{path}{(default: fastp) Path to fastp.}
}
\value{
A logical that say if fastp is installed.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage.
}
\examples{
MiscMetabar::is_fastp_installed()
}
\seealso{
\code{\link[=fastp]{fastp()}}
}
\author{
Adrien Taudière
}
