% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsx.R
\name{is_itsx_installed}
\alias{is_itsx_installed}
\title{Test if ITSx is installed}
\usage{
is_itsx_installed(path = find_itsx(), args_before_itsx = "")
}
\arguments{
\item{path}{(default: \code{\link[=find_itsx]{find_itsx()}}) Path to ITSx.}

\item{args_before_itsx}{(String, default "") A one line bash command run
before ITSx, e.g. the value returned by \code{\link[=install_itsx]{install_itsx()}}:
\code{"source ~/miniforge3/etc/profile.d/conda.sh && conda activate itsxenv && "}.}
}
\value{
A logical that says if ITSx can be run.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage.
}
\examples{
MiscMetabar::is_itsx_installed()
}
\seealso{
\code{\link[=find_itsx]{find_itsx()}}, \code{\link[=install_itsx]{install_itsx()}}, \code{\link[=itsx_pq]{itsx_pq()}}
}
\author{
Adrien Taudière
}
