% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmseqs2.R
\name{is_mmseqs2_installed}
\alias{is_mmseqs2_installed}
\title{Check whether MMseqs2 is installed and callable}
\usage{
is_mmseqs2_installed(path = find_mmseqs2())
}
\arguments{
\item{path}{Path to the \code{mmseqs} binary (default: \code{\link[=find_mmseqs2]{find_mmseqs2()}}).}
}
\value{
Logical.
}
\description{
Tries to run \verb{mmseqs version} and returns \code{TRUE} if it succeeds.
}
\examples{
is_mmseqs2_installed()
}
\seealso{
\code{\link[=find_mmseqs2]{find_mmseqs2()}}, \code{\link[=install_mmseqs2]{install_mmseqs2()}}, \code{\link[=assign_mmseqs2]{assign_mmseqs2()}}
}
\author{
Adrien Taudière
}
