% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_multiqc_installed}
\alias{is_multiqc_installed}
\title{Test if MultiQC is installed.}
\usage{
is_multiqc_installed(path = "multiqc")
}
\arguments{
\item{path}{(default: multiqc) Path to MultiQC}
}
\value{
A logical that say if MultiQC is installed.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stable"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_multiqc_installed()
}
\author{
Adrien Taudière
}
