% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_mumu_installed}
\alias{is_mumu_installed}
\title{Test if mumu is installed.}
\usage{
is_mumu_installed(path = "mumu")
}
\arguments{
\item{path}{(default: mumu) Path to mumu}
}
\value{
A logical that say if mumu is install in
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_mumu_installed()
}
\author{
Adrien Taudière
}
