% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_swarm_installed}
\alias{is_swarm_installed}
\title{Test if swarm is installed.}
\usage{
is_swarm_installed(path = "swarm")
}
\arguments{
\item{path}{(default: swarm) Path to falco}
}
\value{
A logical that say if swarm is install in
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_swarm_installed()
}
\author{
Adrien Taudière
}
