% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_test.R
\name{lefser_pq}
\alias{lefser_pq}
\title{Run LEfSe on a phyloseq object}
\usage{
lefser_pq(
  physeq,
  bifactor = NULL,
  modalities = NULL,
  compute_relativeAb = TRUE,
  by_clade = FALSE,
  ...
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{bifactor}{(required) The name of a column present in the \verb{@sam_data} slot
of the physeq object. Must be a character vector or a factor.}

\item{modalities}{(default NULL) A vector of modalities to keep in the analysis.
If NULL, all modalities present in bifactor are kept. Note that only two
modalities are allowed.
@param compute_relativeAb (logical, default TRUE) Do we compute relative abundance
before running LEfSe?
@param by_clade (logical, default FALSE) Do we use the lefserClades function
(which test for different depth in the taxonomic classification) or the
lefser function (taxa-level)?
@param ... Additional arguments passed on to \code{lefser::lefser()}}
}
\value{
The result of lefser::lefser() or lefser::lefserClades()
}
\description{
Run LEfSe on a phyloseq object
}
\details{
It is a wrapper of the \code{lefser::lefser()} and \code{lefser::lefserClades()} functions.\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}
}
\examples{
res_lefse <- lefser_pq(data_fungi,
  bifactor = "Height",
  modalities = c("Low", "High")
)
lefser::lefserPlot(res_lefse)
}
\author{
Adrien Taudière
}
