% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{lulu_phyloseq}
\alias{lulu_phyloseq}
\title{(Lulu)[https://github.com/adrientaudiere/lulu] reclustering of class `physeq` (forked from \url{https://github.com/tobiasgf/lulu}).
 See https://www.nature.com/articles/s41467-017-01312-x for more information on the method.
`r lifecycle::badge("experimental")`}
\usage{
lulu_phyloseq(
  physeq,
  nproc = 1,
  id = 0.84,
  vsearchpath = "vsearch",
  verbose = FALSE,
  clean_physeq = FALSE
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{nproc}{(default 1)
Set to number of cpus/processors to use for the clustering}

\item{id}{(default = 0.84): id for --usearch_global}

\item{vsearchpath}{(default = "vsearch"): path to vsearch}

\item{verbose}{(Logical): if true, print some additional messages}

\item{clean_physeq}{(Logical, default = FALSE) If true, empty samples and empty ASV are discarded before clustering.}
}
\value{
a list of for object
- "new_physeq": The new phyloseq object (class physeq)
- "discrepancy_vector": A vector of discrepancy showing for each taxonomic level the proportion of identic value 
  before and after lulu reclustering. A value of 0.6 stands for 60% of ASV before re-clusering have 
  identical value after re-clustering. In other word, 40% of ASV are assigned to a different taxonomic 
  value. NA value are not counted as discrepancy.
- "res_lulu": A list of the result from the lulu function 
- "merged_ASV": the data.frame used to merged ASV
}
\description{
(Lulu)[https://github.com/adrientaudiere/lulu] reclustering of class `physeq` (forked from \url{https://github.com/tobiasgf/lulu}).
 See https://www.nature.com/articles/s41467-017-01312-x for more information on the method.
`r lifecycle::badge("experimental")`
}
\examples{
data(data_fungi_sp_known)
lulu_phyloseq(data_fungi_sp_known)
}
\references{
- LULU : \url{https://github.com/adrientaudiere/lulu}
 forked from \url{https://github.com/tobiasgf/lulu}.
- VSEARCH can be downloaded from
 \url{https://github.com/torognes/vsearch}.
}
\author{
Adrien Taudière \email{adrien.taudiere@zaclys.net}
}
