% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{merge_krona}
\alias{merge_krona}
\title{Merge Krona files using \href{https://github.com/marbl/Krona/wiki}{KronaTools}.}
\usage{
merge_krona(files = NULL, output = "mergeKrona.html")
}
\arguments{
\item{files}{(required): path to html files to merged}

\item{output}{(default="mergeKrona.html") : path to the output files}
}
\value{
A html file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Need the installation of kronatools on the computer (\href{https://github.com/marbl/Krona/wiki/Installing}{installation instruction}).
Function merge_krona allows merging multiple html files in one interactive krona file
}
\examples{
data("GlobalPatterns")
GA <- subset_taxa(GlobalPatterns, Phylum == "Acidobacteria")
# krona(GA, "Number.of.sequences.html")
# krona(GA, "Number.of.OTUs.html", nb_seq = F)
# merge_krona(c("Number.of.sequences.html", "Number.of.OTUs.html"))
}
\seealso{
\code{\link{krona}}
}
\author{
Adrien Taudière
}
