% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{MiscMetabar-deprecated}
\alias{MiscMetabar-deprecated}
\alias{physeq_graph_test}
\alias{adonis_phyloseq}
\alias{clean_physeq}
\alias{lulu_phyloseq}
\alias{otu_circle}
\alias{biplot_physeq}
\alias{read_phyloseq}
\alias{write_phyloseq}
\alias{sankey_phyloseq}
\alias{summary_plot_phyloseq}
\alias{plot_deseq2_phyloseq}
\alias{plot_edgeR_phyloseq}
\alias{venn_phyloseq}
\alias{ggVenn_phyloseq}
\alias{hill_tuckey_phyloseq}
\alias{hill_phyloseq}
\alias{physeq_heat_tree}
\title{Deprecated function(s) in the MiscMetabar package}
\usage{
physeq_graph_test(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the MiscMetabar package. They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
\link{graph_test_pq} \tab now a synonym for \link{physeq_graph_test}\cr
\link{adonis_pq} \tab now a synonym for \link{adonis_phyloseq}\cr
\link{clean_pq} \tab now a synonym for \link{clean_physeq}\cr
\link{lulu_pq} \tab now a synonym for \link{lulu_phyloseq}\cr
\link{circle_pq} \tab now a synonym for \link{otu_circle}\cr
\link{biplot_pq} \tab now a synonym for \link{biplot_physeq}\cr
\link{read_pq} \tab now a synonym for \link{read_phyloseq}\cr
\link{write_pq} \tab now a synonym for \link{write_phyloseq}\cr
\link{sankey_pq} \tab now a synonym for \link{sankey_phyloseq}\cr
\link{summary_plot_pq} \tab now a synonym for \link{summary_plot_phyloseq}\cr
\link{plot_edgeR_pq} \tab now a synonym for \link{plot_edgeR_phyloseq}\cr
\link{plot_deseq2_pq} \tab now a synonym for \link{plot_deseq2_phyloseq}\cr
\link{venn_pq} \tab now a synonym for \link{venn_phyloseq}\cr
\link{ggvenn_pq} \tab now a synonym for \link{ggVenn_phyloseq}\cr
\link{hill_tuckey_pq} \tab now a synonym for \link{hill_tuckey_phyloseq}\cr
\link{hill_pq} \tab now a synonym for \link{hill_phyloseq}\cr
\link{heat_tree_pq} \tab now a synonym for \link{physeq_heat_tree}\cr
}
}

