% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscMetabar-package.R
\docType{package}
\name{MiscMetabar-package}
\alias{MiscMetabar-package}
\alias{MiscMetabar}
\title{\code{MiscMetabar} package}
\description{
Functions to help analyze and visualize metabarcoding data. Mainly based on
the phyloseq and dada2 packages.

The MiscMetabar package aims to facilitate the description, transformation, exploration, and reproducibility of metabarcoding analysis. Mainly build on the top of phyloseq, dada2 R packages. MiscMetabar help to build reproducible and robust bioinformatic pipeline in R. MiscMetabar make ecological analysis of alpha and beta-diversity simple and powerfull by integrating a large number of analysis, some of them from other R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adrientaudiere/MiscMetabar}
  \item \url{https://adrientaudiere.github.io/MiscMetabar/}
  \item Report bugs at \url{https://github.com/adrientaudiere/MiscMetabar/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

}
\keyword{internal}
