% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscMetabar-package.R
\docType{package}
\name{MiscMetabar}
\alias{MiscMetabar}
\title{\code{MiscMetabar} package}
\description{
Functions to help analyse and visualise metabarcoding data. Mainly based on
the phyloseq and dada2 packages.
}
