% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_misc.R
\name{MM_idtaxa}
\alias{MM_idtaxa}
\title{A wrapper of \code{\link[DECIPHER:IdTaxa]{DECIPHER::IdTaxa()}}}
\usage{
MM_idtaxa(
  test,
  trainingSet,
  column_names = c("Kingdom_idtaxa", "Phyla_idtaxa", "Class_idtaxa", "Order_idtaxa",
    "Family_idtaxa", "Genus_idtaxa", "Species_idtaxa", "VT_idtaxa"),
  ...
)
}
\arguments{
\item{test}{(required)}

\item{trainingSet}{(required)}

\item{column_names}{(optional but often needed): names for the
column of the resulted tibble}

\item{...}{Additional arguments passed on to \code{\link[DECIPHER]{IdTaxa}}}
}
\value{
a tibble that can be used for tax_table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Adrien Taudière
}
