% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmseqs2.R
\name{mmseqs2_clustering}
\alias{mmseqs2_clustering}
\title{Recluster sequences of a phyloseq object or cluster a list of DNA sequences
using MMseqs2 software}
\usage{
mmseqs2_clustering(
  physeq = NULL,
  dna_seq = NULL,
  nproc = 1,
  id = 0.97,
  mmseqs2path = find_mmseqs2(),
  tax_adjust = 0,
  rank_propagation = FALSE,
  mmseqs2_cluster_method = c("easy-cluster", "easy-linclust"),
  coverage = 0.8,
  cov_mode = 0,
  cluster_mode = 0,
  mmseqs2_args = "",
  keep_temporary_files = FALSE
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{dna_seq}{You may directly use a character vector of DNA sequences
in place of physeq args. When physeq is set, dna sequences take the
value of \code{physeq@refseq}.}

\item{nproc}{(default: 1) Number of threads.}

\item{id}{(default: 0.97) Minimum sequence identity threshold (0--1).}

\item{mmseqs2path}{Path to the \code{mmseqs} binary
(default: \code{\link[=find_mmseqs2]{find_mmseqs2()}}).}

\item{tax_adjust}{(Default 0) See the man page
of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}} for more details.
To conserve the taxonomic rank of the most abundant taxa (ASV, OTU, ...),
set \code{tax_adjust} to 0 (default).}

\item{rank_propagation}{(logical, default FALSE). Do we propagate the
NA value from lower taxonomic rank to upper rank?
See the man page of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}} for more details.}

\item{mmseqs2_cluster_method}{(default: \code{"easy-cluster"}) Either
\code{"easy-cluster"} (cascaded clustering, more sensitive) or
\code{"easy-linclust"} (linear-time clustering, faster for huge datasets).}

\item{coverage}{(numeric, default: 0.8) Alignment coverage threshold
(0--1), passed to \code{-c}.}

\item{cov_mode}{(integer, default: 0) Coverage mode:
\itemize{
\item \code{0}: \code{alnRes / max(qLen, tLen)}
\item \code{1}: \code{alnRes / tLen}
\item \code{2}: \code{alnRes / qLen}
}}

\item{cluster_mode}{(integer, default: 0) Clustering algorithm:
\itemize{
\item \code{0}: greedy set cover (default)
\item \code{1}: connected components
\item \code{2}: greedy incremental
}}

\item{mmseqs2_args}{(character, default: \code{""}) Additional arguments
passed to the MMseqs2 clustering command.}

\item{keep_temporary_files}{(logical, default: FALSE) Keep intermediate
files for debugging?}
}
\value{
A new object of class \code{physeq} or a data.frame of cluster
membership if \code{dna_seq} was used.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


A wrapper of the \href{https://github.com/soedinglab/MMseqs2}{MMseqs2}
\code{easy-cluster} or \code{easy-linclust} workflow.
}
\details{
This function is mainly a wrapper of the work of others.
Please cite \href{https://github.com/soedinglab/MMseqs2}{MMseqs2}.
}
\examples{
\dontshow{if (MiscMetabar::is_mmseqs2_installed()) withAutoprint(\{ # examplesIf}
\donttest{
d_mm <- mmseqs2_clustering(data_fungi_mini)
}
\dontshow{\}) # examplesIf}
}
\references{
MMseqs2 can be downloaded from
\url{https://github.com/soedinglab/MMseqs2}.
More information in the associated publication
\doi{10.1038/nbt.3988}.
}
\seealso{
\code{\link[=postcluster_pq]{postcluster_pq()}}, \code{\link[=vsearch_clustering]{vsearch_clustering()}}, \code{\link[=swarm_clustering]{swarm_clustering()}}
}
\author{
Adrien Taudière
}
