% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{mumu_pq}
\alias{mumu_pq}
\title{MUMU reclustering of class \code{physeq}}
\usage{
mumu_pq(
  physeq,
  nproc = 1,
  id = 0.84,
  vsearchpath = "vsearch",
  mumupath = "mumu",
  lulu_exact = FALSE,
  verbose = FALSE,
  clean_pq = TRUE,
  keep_temporary_files = FALSE,
  extra_mumu_args = NULL
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{nproc}{(default 1)
Set to number of cpus/processors to use for the clustering}

\item{id}{(default: 0.84) id for --usearch_global.}

\item{vsearchpath}{(default: vsearch) path to vsearch.}

\item{mumupath}{path to mumu. See \href{https://github.com/frederic-mahe/mumu}{mumu}
for installation instruction}

\item{lulu_exact}{(logical) If true, use the exact same algorithm as LULU
corresponding to the --legacy option of mumu. Need mumu version >= v1.1.0}

\item{verbose}{(logical) If true, print some additional messages.}

\item{clean_pq}{(logical) If true, empty samples and empty ASV are discarded
before clustering.}

\item{keep_temporary_files}{(logical, default: FALSE) Do we keep temporary files}

\item{extra_mumu_args}{(character, default: NULL) Additional arguments passed
on to mumu command line. See \verb{man mumu} into bash for details. Major args are
\code{--minimum_match}, \code{--minimum_ratio_type}, \code{--minimum_ratio},
\code{--minimum_relative_cooccurence} and \code{--threads}}
}
\value{
a list of for object
\itemize{
\item "new_physeq": The new phyloseq object (class physeq)
\item "mumu_results": The log file of the mumu software. Run \verb{man mumu} into
bash to obtain details about columns' signification.
}
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


See https://www.nature.com/articles/s41467-017-01312-x for more information
on the original method LULU. This is a wrapper of
\href{https://github.com/frederic-mahe/mumu}{mumu} a C++ re-implementation
of LULU by Frédéric Mahé
}
\details{
This function is mainly a wrapper of the work of others.
Please cite \href{https://github.com/frederic-mahe/mumu/blob/main/CITATION.cff}{mumu} and
\href{https://www.nature.com/articles/s41467-017-01312-x}{lulu} if you use this function
for your work.
}
\examples{
\dontshow{if (MiscMetabar::is_mumu_installed()) withAutoprint(\{ # examplesIf}
\dontrun{
ntaxa(data_fungi_sp_known)
ntaxa(mumu_pq(data_fungi_sp_known)$new_physeq)
ntaxa(mumu_pq(data_fungi_sp_known, extra_mumu_args = "--minimum_match 90")$new_physeq)
}
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item MUMU: \url{https://github.com/frederic-mahe/mumu}
\item VSEARCH can be downloaded from
\url{https://github.com/torognes/vsearch}.
}
}
\seealso{
\code{\link[=lulu_pq]{lulu_pq()}}
}
\author{
Frédéric Mahé
& Adrien Taudière \email{adrien.taudiere@zaclys.net}
}
