% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{no_legend}
\alias{no_legend}
\title{Discard legend in ggplot2}
\usage{
no_legend()
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stable"></a>
}}


A more memorable shortcut for theme(legend.position = "none").
}
\examples{
plot_refseq_pq(data_fungi)
plot_refseq_pq(data_fungi) + no_legend()
}
\author{
Adrien Taudière
}
