% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_test.R
\name{physeq_graph_test}
\alias{physeq_graph_test}
\title{A wrapper of \code{\link[phyloseqGraphTest]{graph_perm_test}} for quick plot with
important statistics}
\usage{
physeq_graph_test(
  physeq,
  fact,
  merge_sample_by = NULL,
  nperm = 999,
  return_plot = TRUE,
  title = "Graph Test",
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(required): Name of the factor to cluster samples by modalities.
Need to be in \code{physeq@sam_data}. This should be a factor.
with two or more levels.}

\item{merge_sample_by}{(default : NULL) : a vector to determine
which samples to merge using the \code{\link[speedyseq]{merge_samples2}} function.
Need to be in \code{physeq@sam_data}}

\item{nperm}{(default = 999): The number of permutations to perform.}

\item{return_plot}{(default = TRUE) : do we return only the result
of the test or do we plot the result}

\item{...}{other params for be passed on
\code{phyloseqGraphTest::graph_perm_test} function}
}
\value{
a ggplot with a subtitle indicating the pvalue
and the number of permutations
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Adrien Taudière
}
