% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deseq2_edgeR.R
\name{plot_edgeR_pq}
\alias{plot_edgeR_pq}
\title{Plot edgeR results for a phyloseq or a edgeR object.}
\usage{
plot_edgeR_pq(
  physeq,
  contrast = NULL,
  alpha = 0.01,
  taxolev = "Genus",
  color_tax = "Phylum",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{contrast}{(required):This argument specifies what comparison
to extract from the object to build a results table.
See \code{\link[DESeq2]{results}} man page for more details.}

\item{alpha}{(default: 0.01): the significance cutoff used for optimizing
the independent filtering. If the adjusted p-value cutoff (FDR) will be a
value other than 0.1, alpha should be set to that value.}

\item{taxolev}{taxonomic level of interest}

\item{color_tax}{taxonomic level used for
color assignation}

\item{verbose}{(logical): whether the function print some
information during the computation}

\item{...}{Additional arguments passed on to \code{\link[edgeR]{exactTest}}
or \code{\link[ggplot2]{ggplot}}}
}
\value{
A \code{\link{ggplot}}2 plot representing edgeR results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
\dontrun{
data(GlobalPatterns)
plot_edgeR_pq(GlobalPatterns, c("SampleType", "Soil", "Feces"),
  color_tax = "Kingdom"
)
plot_edgeR_pq(GlobalPatterns, c("SampleType", "Soil", "Feces"),
  taxolev = "Class", color_tax = "Kingdom"
)
}
}
\seealso{
\code{\link[edgeR]{exactTest}}

\code{\link{plot_deseq2_pq}}
}
\author{
Adrien Taudière
}
