% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsearch_wrappers.R
\name{plot_ee_rate_dist}
\alias{plot_ee_rate_dist}
\title{Plot the distribution of expected error (EE) rates of reads}
\usage{
plot_ee_rate_dist(
  fastq_input,
  n_bins = 30,
  plot_title = "Distribution of the expected error (EE) rate of reads"
)
}
\arguments{
\item{fastq_input}{(required) Path to a fastq file (optionally gzipped).}

\item{n_bins}{(integer, default 30) Number of histogram bins.}

\item{plot_title}{(character) Title of the plot.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Wrapper of \code{\link[Rsearch:plot_ee_rate_dist]{Rsearch::plot_ee_rate_dist()}}. Plot the distribution of the
expected error (EE) rate of the reads of a fastq file, a key diagnostic
before quality filtering (e.g. with \code{maxEE} thresholds).
}
\examples{
\dontshow{if (requireNamespace("Rsearch", quietly = TRUE)) withAutoprint(\{ # examplesIf}
plot_ee_rate_dist(
  system.file("extdata", "ex.fastq", package = "MiscMetabar")
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[Rsearch:plot_ee_rate_dist]{Rsearch::plot_ee_rate_dist()}}, \code{\link[=plot_read_quality]{plot_read_quality()}}
}
\author{
Adrien Taudière
}
