% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mt}
\alias{plot_mt}
\title{Plot the result of a mt test (\code{\link[phyloseq]{mt}})}
\usage{
plot_mt(mt = NULL, alpha = 0.05, color_tax = "Class", taxa = "Species")
}
\arguments{
\item{mt}{(required): result of a mt test}

\item{alpha}{(default = 0.05): Choose the cut off p-value to plot taxa}

\item{color_tax}{: A taxonomic level to color the points}

\item{taxa}{: The taxonomic level choose for x-positioning
data("GlobalPatterns")
res <- mt(GlobalPatterns, "SampleType", test="f")
plot_mt(res, color_tax = "Phylum") + scale_color_hue()}
}
\value{
a \code{\link{ggplot}}2 plot of result of a mt test
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\seealso{
\code{\link[phyloseq]{mt}}
}
\author{
Adrien Taudière
}
