% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_tsne_pq}
\alias{plot_tsne_pq}
\title{Plot a tsne low dimensional representation of a phyloseq object}
\usage{
plot_tsne_pq(
  physeq,
  method = "bray",
  dims = 2,
  theta = 0,
  perplexity = 30,
  fact = NA,
  ellipse_level = 0.95,
  plot_dims = c(1, 2),
  na_remove = TRUE,
  force_factor = TRUE,
  ...
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{method}{A method to calculate distance using \code{vegan::vegdist()} function (default: "bray")}

\item{dims}{(Int) Output dimensionality (default: 2)}

\item{theta}{(Numeric) Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.0 see details in the man page of \code{Rtsne::Rtsne}).}

\item{perplexity}{(Numeric) Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details in the man page of \code{Rtsne::Rtsne})}

\item{fact}{Name of the column in \code{physeq@sam_data} used to color points and compute ellipses.}

\item{ellipse_level}{The level used in stat_ellipse. Set to NULL to discard ellipse (default = 0.95)}

\item{plot_dims}{A vector of 2 values defining the rank of dimension to plot (default: c(1,2))}

\item{na_remove}{(logical, default TRUE) Does the samples with NA values in fact are removed? (default: true)}

\item{force_factor}{(logical, default TRUE) Force the fact column to be a factor.}

\item{...}{Additional arguments passed on to \code{Rtsne::Rtsne()}}
}
\value{
A ggplot object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Partially inspired by \code{phylosmith::tsne_phyloseq()} function developed by Schuyler D. Smith.
}
\details{
As explained in the documentation of \code{\link[=tsne_pq]{tsne_pq()}}, t-SNE is a \strong{local}
technique: the plot is reliable for \emph{neighbourhood}, \emph{outlier} and \emph{cluster
identification}, but the distances between points or clusters, the cluster
density and the class separability it suggests must not be read as faithful
to the original space (Jeon et al., 2026). See \code{\link[=tsne_pq]{tsne_pq()}} for details and
for global alternatives (PCA, MDS/PCoA).
}
\examples{
\dontshow{if (tolower(Sys.info()[["sysname"]]) != "windows") withAutoprint(\{ # examplesIf}
if (requireNamespace("Rtsne")) {
  plot_tsne_pq(data_fungi_mini, fact = "Height", perplexity = 15)
}
\donttest{
if (requireNamespace("Rtsne")) {
  plot_tsne_pq(data_fungi_mini, fact = "Time") +
    geom_label(aes(label = Sample_id, fill = Time))
  plot_tsne_pq(data_fungi_mini,
    fact = "Time", na_remove = FALSE,
    force_factor = FALSE
  )
}
}
\dontshow{\}) # examplesIf}
}
\references{
Jeon, H., Park, J., Shin, S., & Seo, J. (2026). Stop Misusing t-SNE and UMAP
for Visual Analytics. \emph{IEEE Transactions on Visualization and Computer
Graphics}. \doi{10.48550/arXiv.2506.08725}

van der Maaten, L., & Hinton, G. (2008). Visualizing Data using t-SNE.
\emph{Journal of Machine Learning Research}, 9, 2579-2605.
}
\author{
Adrien Taudière
}
