% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{psmelt_samples_pq}
\alias{psmelt_samples_pq}
\title{Build a sample information tibble from physeq object}
\usage{
psmelt_samples_pq(
  physeq,
  q = c(0, 1, 2),
  hill_scales = lifecycle::deprecated(),
  filter_zero = TRUE,
  rarefy_by_sample = FALSE,
  rngseed = FALSE,
  verbose = TRUE,
  taxa_ranks = NULL,
  ...
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{q}{(numeric vector) Hill diversity orders to compute. If NULL, no
Hill numbers are computed. Default computes Hill number 0 (species
richness), 1 (exponential of Shannon index) and 2 (inverse of Simpson
index). Formerly \code{q}. Hill numbers are more appropriate in DNA
metabarcoding studies when \code{q > 0} (Alberdi & Gilbert, 2019;
Calderón-Sanou et al., 2019).}

\item{hill_scales}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{q} instead.}

\item{filter_zero}{(logical, default TRUE) Do we filter non present OTU from
samples ? For the moment, this has no effect on the result because the dataframe
is grouped by samples with abundance summed across OTU.}

\item{rarefy_by_sample}{(logical, default FALSE) If TRUE, rarefy
samples using \code{\link[phyloseq:rarefy_even_depth]{phyloseq::rarefy_even_depth()}} function.}

\item{rngseed}{(Optional). A single integer value passed to
\code{\link[phyloseq:rarefy_even_depth]{phyloseq::rarefy_even_depth()}}, which is used to fix a seed for
reproducibly random number generation (in this case, reproducibly
random subsampling). If set to FALSE, then no fiddling with the RNG seed
is performed, and it is up to the user to appropriately call set.seed
beforehand to achieve reproducible results. Default is FALSE.}

\item{verbose}{(logical). If TRUE, print additional information.}

\item{taxa_ranks}{A vector of taxonomic ranks. For examples c("Family","Genus").
If taxa ranks is not set (default value = NULL), taxonomic information are not
present in the resulting tibble.}

\item{...}{Additional arguments passed to \code{\link[=divent_hill_matrix_pq]{divent_hill_matrix_pq()}} and
hence to \code{\link[divent:div_hill]{divent::div_hill()}} (e.g. \code{estimator = "naive"}). Only used
when \code{q} is not NULL.}
}
\value{
A tibble with a row for each sample. Columns provide information
from \code{sam_data} slot as well as hill numbers, Abundance (nb of sequences),
and Abundance_log10 (\emph{log10(1+Abundance)}).
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Hill numbers are the number of equiprobable species giving the same diversity
value as the observed distribution.

Note that contrary to \code{\link[=hill_pq]{hill_pq()}}, this function does not take into
account for difference in the number of sequences per samples/modalities.
You may use rarefy_by_sample = TRUE if the mean number of sequences per
samples differs among modalities.
}
\examples{
if (requireNamespace("ggstatsplot")) {
  psm_tib <- psmelt_samples_pq(data_fungi_mini, hill_scales = c(0, 2, 7))
  ggstatsplot::ggbetweenstats(psm_tib, Height, Hill_0)
}
\donttest{
if (requireNamespace("ggstatsplot")) {
  ggstatsplot::ggbetweenstats(psm_tib, Height, Hill_7)

  psm_tib_tax <- psmelt_samples_pq(data_fungi_mini, taxa_ranks = c("Class", "Family"))
  ggplot(filter(psm_tib_tax, Abundance > 2000), aes(y = Family, x = Abundance, fill = Time)) +
    geom_bar(stat = "identity") +
    facet_wrap(~Height)
}
}
}
\references{
Alberdi, A., & Gilbert, M. T. P. (2019). A guide to the application of
Hill numbers to DNA-based diversity analyses. \emph{Molecular Ecology Resources}.
\doi{10.1111/1755-0998.13014}

Calderón-Sanou, I., Münkemüller, T., Boyer, F., Zinger, L., & Thuiller, W.
(2019). From environmental DNA sequences to ecological conclusions: How
strong is the influence of methodological choices? \emph{Journal of Biogeography},
47. \doi{10.1111/jbi.13681}
}
\author{
Adrien Taudière
}
