% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{read_phyloseq}
\alias{read_phyloseq}
\title{Read phyloseq object from multiple csv tables and a phylogenetic tree
in Newick format.}
\usage{
read_phyloseq(path = NULL, taxa_are_rows = FALSE)
}
\arguments{
\item{path}{(required) : a path to the folder to read the phyloseq object}

\item{taxa_are_rows}{(required, default to FALSE) : see ?phyloseq for details}
}
\value{
One to four csv tables (refseq.csv, otu_table.csv, tax_table.csv, sam_data.csv)
and if present a phy_tree in Newick format. At least the otu_table.csv need to be present.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
\dontrun{
read_phyloseq(path = "phyloseq_data")
}

}
