% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedyseq_functions.R
\name{sample_data_stable}
\alias{sample_data_stable}
\title{Create sample data without adjusting row/sample names}
\usage{
sample_data_stable(object)
}
\arguments{
\item{object}{A "data.frame"-class object}
}
\description{
\code{phyloseq::sample_data()} will change the sample names from the row names if
they are \code{as.character(seq(1, row(object)))}. This function instead keeps the
names as is.
}
\examples{
x <- data.frame(var1 = letters[1:3], var2 = 7:9)
rownames(x)
sample_data(x)
sample_data_stable(x)
}
\author{
Michael R. McLaren (orcid: \href{https://orcid.org/0000-0003-1575-473X}{0000-0003-1575-473X})
}
\keyword{internal}
