% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_misc.R
\name{sample_data_with_new_names}
\alias{sample_data_with_new_names}
\title{Load sample data from file and rename samples using names of samples and an
optional order}
\usage{
sample_data_with_new_names(file, names_of_samples, samples_order, ...)
}
\arguments{
\item{names_of_samples}{(required) a vector of sample names}

\item{samples_order}{Optional numeric vector to sort sample names}

\item{...}{Other arguments passed on to \code{\link[utils:read.table]{utils::read.delim()}} function.}

\item{file_path}{(required) a path to the sample_data file}
}
\value{
A data.frame from file_path and new names
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\seealso{
\code{\link[=rename_samples]{rename_samples()}}
}
\author{
Adrien Taudière
}
