% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{search_exact_seq_pq}
\alias{search_exact_seq_pq}
\title{Search for exact matching of sequences using complement, reverse and reverse-complement}
\usage{
search_exact_seq_pq(physeq, sequences)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
A list of data-frames for each input sequences with the name, the sequences and the
number of occurences of the original sequence, the complement sequence,
the reverse sequence and the reverse-complement sequence.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data("data_fungi")
}
\author{
Adrien Taudière
}
