% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{select_one_sample}
\alias{select_one_sample}
\title{Select one sample from a physeq object}
\usage{
select_one_sample(physeq, sam_name, silent = FALSE)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{sam_name}{(required) The sample name to select}

\item{silent}{(logical) If true, no message are printing.}
}
\value{
A new \code{\link{phyloseq-class}} object with one sample
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Mostly for internal used, for example in function \code{\link[=track_wkflow_samples]{track_wkflow_samples()}}.
}
\examples{

A8_005 <- select_one_sample(data_fungi, "A8-005_S4_MERGED.fastq.gz")
A8_005
}
\author{
Adrien Taudière
}
