% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pq.R
\name{srs_pq}
\alias{srs_pq}
\title{Scaling with Ranked Subsampling (SRS) normalization of a phyloseq object}
\usage{
srs_pq(physeq, Cmin = NULL, ...)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{Cmin}{(integer) the common scaling depth. Defaults to
\code{min(sample_sums(physeq))}.}

\item{...}{Additional arguments passed on to \code{\link[SRS:SRS]{SRS::SRS()}}.}
}
\value{
A new \code{\link[phyloseq]{phyloseq-class}} object with the SRS
normalised \code{otu_table}.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Wrapper around \code{\link[SRS:SRS]{SRS::SRS()}} (Heidrich et al. 2021,
\doi{10.7717/peerj.9593}) which scales all samples to a common count
\code{Cmin} while preserving the rank order of OTU abundances.
}
\examples{
\dontshow{if (rlang::is_installed("SRS")) withAutoprint(\{ # examplesIf}
data_f_srs <- srs_pq(data_fungi_mini)
sample_sums(data_f_srs)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[SRS:SRS]{SRS::SRS()}}, \code{\link[=rarefy_pq]{rarefy_pq()}}
}
\author{
Adrien Taudière
}
