% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{subset_taxa_pq}
\alias{subset_taxa_pq}
\title{Subset taxa using a conditional named boolean vector.}
\usage{
subset_taxa_pq(physeq, condition, verbose = TRUE, clean_pq = TRUE)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{condition}{A named boolean vector to subset taxa. Length must fit
the number of taxa and names must match taxa_names}

\item{verbose}{(logical) Informations are printed}

\item{clean_pq}{(logical)
If set to TRUE, empty samples are discarded after subsetting ASV}
}
\description{
Subset taxa using a conditional named boolean vector.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The main objective of this function is to complete the \code{\link[phyloseq:subset_taxa-methods]{phyloseq::subset_taxa()}}
function by propose a more easy way of subset_taxa using a named boolean vector.
Names must match taxa_names.
}
