% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{subset_taxa_random}
\alias{subset_taxa_random}
\title{Randomly subset a specified number of taxa from a phyloseq object}
\usage{
subset_taxa_random(physeq, nb_taxa, seed = NULL, clean_pq = TRUE, verbose = TRUE)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{nb_taxa}{(integer) The number of taxa to randomly select. Must be
less than or equal to the total number of taxa in the phyloseq object.}

\item{seed}{(integer, optional) Random seed for reproducibility. If NULL
(default), no seed is set.}

\item{clean_pq}{(logical, default TRUE) If set to TRUE, empty samples are
discarded after subsetting taxa.}

\item{verbose}{(logical, default TRUE) If TRUE, informative messages are
printed.}
}
\value{
A new phyloseq object with the randomly selected taxa
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


A simple helper function to randomly subset a specified number of taxa
from a phyloseq object. This is useful for testing, creating example
datasets, or for computational efficiency when working with large datasets.
}
\examples{
# Randomly select 20 taxa
subset_taxa_random(data_fungi, nb_taxa = 20)

# With a seed for reproducibility
subset_taxa_random(data_fungi, nb_taxa = 20, seed = 123)

# Most succinct way to get 20 random taxa
subset_taxa_random(data_fungi, 20)

}
\seealso{
\code{\link[=subset_taxa_pq]{subset_taxa_pq()}}, \code{\link[=filt_taxa_pq]{filt_taxa_pq()}}
}
\author{
Adrien Taudière
}
