% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{subset_taxa_taxControl}
\alias{subset_taxa_taxControl}
\title{Subset taxa using a taxa control (e.g. truffle root tips) through 3 methods.}
\usage{
subset_taxa_taxControl(
  physeq,
  taxa_distri,
  method = "mean",
  min_diff_for_cutoff = 10
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{taxa_distri}{(required): a vector of length equal to the number of samples with the number of sequences per samples for the taxa control}

\item{method}{(default = "mean"): a method to calculate the cut-off value. There is 6 methods:
(i) cutoffSeq: discard taxa with less than the number of sequence than taxa control,
(ii) cutoffMixt: using mixture models,
(iii) cutoffDiff: using a minimum difference threshold (need the argument min_diff_for_cutoff)
(iv) min: the minimum of the three firsts methods
(v) max: the maximum of the three firsts methods
(vi) mean: the mean of the three firsts methods}

\item{min_diff_for_cutoff}{(default = 10): argument for method cutoffDiff}
}
\value{
A new \code{\link{phyloseq-class}} object.
}
\description{
Subset taxa using a taxa control (e.g. truffle root tips) through 3 methods.
}
\author{
Adrien Taudière
}
