% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary_plot_phyloseq}
\alias{summary_plot_phyloseq}
\title{Summarise a \code{\link{phyloseq-class}} object using a plot.}
\usage{
summary_plot_phyloseq(
  physeq,
  add_info = TRUE,
  min_seq_samples = 500,
  clean_phyloseq = TRUE
)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}} object.
For the moment refseq slot need to be not Null.}

\item{add_info}{(default TRUE): Does the bottom down corner contain
extra informations.}

\item{min_seq_samples}{(default 500): Used only when add_info is set
to true to print the number of samples with less sequences than
this number.}

\item{clean_phyloseq}{(Bool, default to TRUE): Does the phyloseq
object is cleaned using the \code{\link[MiscMetabar]{clean_physeq}}
function.}
}
\value{
A ggplot2 object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
data(data_fungi)
summary_plot_phyloseq(data_fungi)
}
