% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pq.R
\name{tmm_pq}
\alias{tmm_pq}
\title{Trimmed Mean of M-values (TMM) normalization of a phyloseq object}
\usage{
tmm_pq(physeq, log = FALSE)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{log}{(logical, default \code{FALSE}) if \code{TRUE}, returns \code{log2(cpm + 1)}.}
}
\value{
A new \code{\link[phyloseq]{phyloseq-class}} object with a TMM
normalised \code{otu_table}.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Wrapper around \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} with \code{method = "TMM"}
(Robinson & Oshlack 2010, \doi{10.1186/gb-2010-11-3-r25}). Returns
counts-per-million scaled by the TMM-derived library sizes.
}
\examples{
\dontshow{if (rlang::is_installed("edgeR")) withAutoprint(\{ # examplesIf}
data_f_tmm <- tmm_pq(data_fungi_mini)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}, \code{\link[edgeR:cpm]{edgeR::cpm()}}
}
\author{
Adrien Taudière
}
