% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{transp}
\alias{transp}
\title{Adds transparency to a vector of colors}
\usage{
transp(col, alpha = 0.5)
}
\arguments{
\item{col}{a vector of colors}

\item{alpha}{(default 0.5) a numeric value between 0 and 1 representing the alpha coefficient; 0: total transparency; 1: no transparency.}
}
\value{
a color vector
}
\description{
Adds transparency to a vector of colors
}
\seealso{
The R package RColorBrewer, proposing a nice selection of color palettes. The viridis package, with many excellent palettes
}
\author{
Thibaut Jombart in \code{adegenet} package
}
