% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{vsearch_clustering}
\alias{vsearch_clustering}
\title{Recluster sequences of an object of class \code{physeq}
or cluster a list of DNA sequences using vsearch software}
\usage{
vsearch_clustering(
  physeq = NULL,
  dna_seq = NULL,
  nproc = 1,
  id = 0.97,
  vsearchpath = "vsearch",
  tax_adjust = 0,
  vsearch_cluster_method = "--cluster_size",
  vsearch_args = "--strand both",
  keep_temporary_files = FALSE
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{dna_seq}{You may directly use a character vector of DNA sequences
in place of physeq args. When physeq is set, dna sequences take the value of
\code{physeq@refseq}}

\item{nproc}{(default: 1)
Set to number of cpus/processors to use for the clustering}

\item{id}{(default: 0.97) level of identity to cluster}

\item{vsearchpath}{(default: vsearch) path to vsearch}

\item{tax_adjust}{(Default 0) See the man page
of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}} for more details.
To conserved the taxonomic rank of the most abundant ASV,
set tax_adjust to 0 (default)}

\item{vsearch_cluster_method}{(default: "--cluster_size) See other possible
methods in the \href{https://github.com/torognes/vsearch/releases/download/v2.23.0/vsearch_manual.pdf}{vsearch pdf manual} (e.g. \code{--cluster_size} or \code{--cluster_smallmem})
\itemize{
\item \code{--cluster_fast} : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence length beforehand.
\item \code{--cluster_size} : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence abundance beforehand.
\item \code{--cluster_smallmem} : Clusterize the fasta sequences in filename without automatically modifying their order beforehand. Sequence are expected to be sorted by decreasing sequence length, unless \emph{--usersort} is used
}}

\item{vsearch_args}{(default : "--strand both") a one length character element defining other parameters to
passed on to vsearch.}

\item{keep_temporary_files}{(logical, default: FALSE) Do we keep temporary files ?
\itemize{
\item temp.fasta (refseq in fasta or dna_seq sequences)
\item cluster.fasta (centroid if method = "vsearch")
\item temp.uc (clusters if method = "vsearch")
}}
}
\value{
A new object of class \code{physeq} or a list of cluster if dna_seq
args was used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
This function use the \code{\link[=merge_taxa_vec]{merge_taxa_vec()}} function to
merge taxa into clusters. By default tax_adjust = 0. See the man page
of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}}.

This function is mainly a wrapper of the work of others.
Please cite \href{https://github.com/torognes/vsearch}{vsearch}.
}
\examples{
# data(data_fungi)
summary_plot_pq(data_fungi)
d_vs <- vsearch_clustering(data_fungi)
summary_plot_pq(d_vs)
}
\references{
VSEARCH can be downloaded from
\url{https://github.com/torognes/vsearch}.
More information in the associated publication
\url{https://pubmed.ncbi.nlm.nih.gov/27781170}.
}
\seealso{
\code{\link[=asv2otu]{asv2otu()}}, \code{\link[=swarm_clustering]{swarm_clustering()}}
}
\author{
Adrien Taudière
}
