% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{vsearch_search_global}
\alias{vsearch_search_global}
\title{Search for a list of sequence in a fasta file against physeq reference
sequences using \href{https://github.com/torognes/vsearch}{vsearch}}
\usage{
vsearch_search_global(
  physeq,
  seq2search,
  vsearchpath = "vsearch",
  id = 0.8,
  iddef = 0
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{seq2search}{(required) path to fasta file}

\item{vsearchpath}{path to vsearch}

\item{id}{(default: 0.8) id for the option \code{--usearch_global} of the vsearch software}

\item{iddef}{(default: 0) iddef for the option \code{--usearch_global} of the vsearch software}
}
\value{
A dataframe with uc results (invisible)
}
\description{
\verb{r lifecycle::badge("maturing")}
}
\examples{
\dontrun{
file_dna <- tempfile("dna.fa")
seqinr::write.fasta("GCCCATTAGTATTCTAGTGGGCATGCCTGTTCGAGCGTCATTTTCA
  ACCCTCAAGCCCCTTATTGCTTGGTGTTGGGAGTTTAGCTGGCTTTATAGCGGTTAACTCCCTAAATATACTGGCG",
  file = file_dna, name = "seq1"
)
res <- vsearch_search_global(data_fungi, file_dna)
unlink(file_dna)

res[res$identity != "*", ]

clean_pq(subset_taxa(data_fungi, res$identity != "*"))
}
}
