% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pq.R
\name{vst_pq}
\alias{vst_pq}
\title{Variance Stabilizing Transformation of a phyloseq object (DESeq2)}
\usage{
vst_pq(physeq, blind = TRUE, fitType = "parametric")
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{blind}{(logical, default \code{TRUE}) passed to DESeq2.}

\item{fitType}{(character, default \code{"parametric"}) passed to DESeq2.}
}
\value{
A new \code{\link[phyloseq]{phyloseq-class}} object with a VST
transformed \code{otu_table}.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Wrapper around \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}
(Love, Huber & Anders 2014, \doi{10.1186/s13059-014-0550-8}). Counts
are incremented by 1 to handle zeros before VST is applied.
}
\examples{
\dontshow{if (rlang::is_installed("DESeq2")) withAutoprint(\{ # examplesIf}
data_f_vst <- vst_pq(data_fungi_mini)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}
}
\author{
Adrien Taudière
}
