% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{accu_plot}
\alias{accu_plot}
\title{Plot accumulation curves for \code{\link{phyloseq-class}} object}
\usage{
accu_plot(
  physeq,
  fact = NULL,
  add_nb_seq = TRUE,
  step = NULL,
  by.fact = FALSE,
  ci_col = NULL,
  col = NULL,
  lwd = 3,
  leg = TRUE,
  print_sam_names = FALSE,
  ci = 2,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(required): Name of the factor in physeq@sam_data used to plot
different lines}

\item{add_nb_seq}{(Default: TRUE, logical):
Either plot accumulation curves using sequences or using samples}

\item{step}{(Integer): distance among points calculated to plot lines. A
low value give better plot but is more time consuming.
Only used if add_nb_seq = TRUE.}

\item{by.fact}{(Default: FALSE, logical):
First merge the OTU table by factor to plot only one line by factor}

\item{ci_col}{: Color vector for confidence intervall.
Only use if add_nb_seq = FALSE.
If add_nb_seq = TRUE, you can use ggplot to modify the plot.}

\item{col}{: Color vector for lines. Only use if add_nb_seq = FALSE.
If add_nb_seq = TRUE, you can use ggplot to modify the plot.}

\item{lwd}{(Default: 3): thickness for lines. Only use if add_nb_seq = FALSE.}

\item{leg}{(Default: TRUE, logical): Plot legend or not. Only use if add_nb_seq = FALSE.}

\item{print_sam_names}{(Default: FALSE, logical): Print samples names or not?
Only use if add_nb_seq = TRUE.}

\item{ci}{(Default: 2, integer) : Confidence intervall value used to multiply the
standard error to plot confidence intervall}

\item{...}{Additional arguments passed on to \code{\link{ggplot}}
if add_nb_seq = TRUE or to \code{\link{plot}} if add_nb_seq = FALSE}
}
\value{
A \code{\link{ggplot}}2 plot representing the richness
accumulation plot if add_nb_seq = TRUE, else, if add_nb_seq = FALSE
return a base plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
data("GlobalPatterns")
GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
accu_plot(GP, "SampleType", add_nb_seq = TRUE, by.fact = TRUE)

}
\seealso{
\code{\link[vegan]{specaccum}}
}
\author{
Adrien Taudière
}
