% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{add_dna_to_phyloseq}
\alias{add_dna_to_phyloseq}
\title{Add dna in \code{refseq} slot of a \code{physeq} object using taxa names and renames taxa
using ASV_1, ASV_2, …}
\usage{
add_dna_to_phyloseq(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
A new \code{\link{phyloseq-class}} object with \code{refseq} slot and new
taxa names
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stable"></a>
}}
}
