% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{add_dna_to_phyloseq}
\alias{add_dna_to_phyloseq}
\title{Add dna in \code{refseq} slot of a \code{physeq} object using taxa names and renames taxa
using prefix_taxa_names and number (default Taxa_1, Taxa_2 ...)}
\usage{
add_dna_to_phyloseq(physeq, prefix_taxa_names = "Taxa_")
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{prefix_taxa_names}{(default "Taxa_"): the prefix of taxa names (eg. "ASV_" or "OTU_")}
}
\value{
A new \code{\link[phyloseq]{phyloseq-class}} object with \code{refseq} slot and new
taxa names
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stable"></a>
}}


Useful in targets bioinformatic pipeline.
}
\author{
Adrien Taudière
}
