% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{add_new_taxonomy_pq}
\alias{add_new_taxonomy_pq}
\title{Add new taxonomic rank to a phyloseq object.}
\usage{
add_new_taxonomy_pq(physeq, ref_fasta, suffix = NULL, ...)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{ref_fasta}{(required) A link to a database.
Pass on to \code{dada2::assignTaxonomy}.}

\item{suffix}{(character) The suffix to name the new columns.
If set to NULL (the default), the basename of the file reFasta
is used.}

\item{...}{Others arguments pass on to \code{dada2::assignTaxonomy}.}
}
\value{
A new \code{\link{phyloseq-class}} object with a larger slot tax_table"
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

One of main use of this function is to add taxonomic assignment from
a new database.
}
\author{
Adrien Taudière
}
