% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{all_object_size}
\alias{all_object_size}
\title{List the size of all objects of the GlobalEnv.}
\usage{
all_object_size()
}
\value{
a list of size
}
\description{
#'
\if{html}{\out{<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">}}
<img src="https://img.shields.io/badge/lifecycle-stable-green" alt="lifecycle-stablealt="lifecycle-maturing">\if{html}{\out{</a>}}

Code from https://tolstoy.newcastle.edu.au/R/e6/help/09/01/1121.html
}
