% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{all_object_size}
\alias{all_object_size}
\title{List the size of all object of the GlobalEnv.}
\usage{
all_object_size()
}
\value{
a list of size
}
\description{
`r lifecycle::badge("stable")`

Code from https://tolstoy.newcastle.edu.au/R/e6/help/09/01/1121.html
}
