% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{asv2otu}
\alias{asv2otu}
\title{Recluster sequences of an object of class \code{physeq}
(e.g. OTUs or ASV from dada)}
\usage{
asv2otu(
  physeq = NULL,
  seq_names = NULL,
  nproc = 1,
  method = "clusterize",
  id = 0.97,
  vsearchpath = "vsearch",
  tax_adjust = 0,
  vsearch_cluster_method = "--cluster_size",
  vsearch_args = "--strand both",
  keep_temporary_files = FALSE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{seq_names}{You may directly use a character vector of DNA sequences
in place of physeq args. When physeq is set, dna sequences take the value of
\code{physeq@refseq}}

\item{nproc}{(default: 1)
Set to number of cpus/processors to use for the clustering}

\item{method}{(default: clusterize)
Set the clustering method.
\itemize{
\item \code{clusterize} use the \code{\link[DECIPHER:Clusterize]{DECIPHER::Clusterize()}} fonction,
\item \code{vsearch} use the vsearch software (https://github.com/torognes/vsearch/)
with arguments \code{--cluster_size} by default (see args \code{vsearch_cluster_method})
and \verb{-strand both} (see args \code{vsearch_args})
}}

\item{id}{(default: 0.97) level of identity to cluster}

\item{vsearchpath}{path to vsearch}

\item{tax_adjust}{See the man page
of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}} for more details.
To conserved the taxonomic rank of the most abundant ASV,
set tax_adjust to 0}

\item{vsearch_cluster_method}{(default: "--cluster_size) See other possible
methods in the \href{https://github.com/torognes/vsearch/releases/download/v2.23.0/vsearch_manual.pdf}{vsearch pdf manual} (e.g. \code{--cluster_size} or \code{--cluster_smallmem})
\itemize{
\item \code{--cluster_fast} : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence length beforehand.
\item \code{--cluster_size} : Clusterize the fasta sequences in filename, automatically sort by decreasing sequence abundance beforehand.
\item \code{--cluster_smallmem} : Clusterize the fasta sequences in filename without automatically modifying their order beforehand. Sequence are expected to be sorted by decreasing sequence length, unless \emph{--usersort} is used
}}

\item{vsearch_args}{(default : "--strand both") a one length character element defining other parameters to
passed on to vsearch.}

\item{keep_temporary_files}{(logical, default: FALSE) Do we keep temporary files
\itemize{
\item temp.fasta (refseq in fasta)
\item cluster.fasta (centroid if method = "vsearch")
\item temp.uc (clusters if method = "vsearch")
}}

\item{...}{Others arguments path to \code{\link[DECIPHER:Clusterize]{DECIPHER::Clusterize()}}}
}
\value{
A new object of class \code{physeq} or a list of cluster if seq_names
args was used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
This function use the \code{merge_taxa_vec} function to
merge taxa into clusters. By default tax_adjust = 1L. See the man page
of \code{\link[=merge_taxa_vec]{merge_taxa_vec()}}.
}
\references{
VSEARCH can be downloaded from
\url{https://github.com/torognes/vsearch}.
More information in the associated publication
\url{https://pubmed.ncbi.nlm.nih.gov/27781170}.
}
