% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{biplot_physeq}
\alias{biplot_physeq}
\title{Visualization of two samples for comparison of physeq object}
\usage{
biplot_physeq(
  physeq,
  merge_sample_by = NULL,
  left_cond = c(FALSE, TRUE),
  left_name = NA,
  right_name = NA,
  left_fill = "#4B3E1E",
  left_col = "#f3f2d9",
  right_fill = "#1d2949",
  right_col = "#1d2949",
  log_10 = TRUE,
  nudge_y = 0.3,
  geomLabel = FALSE,
  text_size = 3,
  ...
)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}}
object}

\item{merge_sample_by}{(default : NULL) : a boolean vector to determine
wich samples to merge for a total of 2 samples at the end.}

\item{left_cond}{(default : c(FALSE, TRUE)) : A condition to select the left sample in the form
of a two-size vector of boolean (e.g. c(TRUE, FALSE)).}

\item{left_name}{: Name fo the left sample.}

\item{right_name}{: Name fo the rigth sample.}

\item{left_fill}{: Fill fo the left sample.}

\item{left_col}{: Color fo the left sample.}

\item{right_fill}{: Fill fo the right sample.}

\item{right_col}{: Color fo the right sample.}

\item{log_10}{(default : TRUE) : Does abundancy is log10 transformed ?}

\item{nudge_y}{: A parameter to control the y position of abundancy values.}

\item{geomLabel}{(default FALSE) : if true use the \code{geom_label} function
instead of \code{geom_text} to indicate the numbers of sequences}

\item{text_size}{(default : 3): default size for the number of sequences}

\item{...}{: other arguments for the ggplot function}
}
\value{
A plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\author{
Adrien Taudière
}
