% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{biplot_physeq}
\alias{biplot_physeq}
\title{Visualization of two samples for comparison of physeq object}
\usage{
biplot_physeq(
  physeq,
  fact = NULL,
  merge_sample_by = NULL,
  inverse_side = FALSE,
  left_name = NULL,
  right_name = NULL,
  left_fill = "#4B3E1E",
  left_col = "#f3f2d9",
  right_fill = "#1d2949",
  right_col = "#1d2949",
  log_10 = TRUE,
  nudge_y = 0.3,
  geomLabel = FALSE,
  text_size = 3,
  ...
)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}}
object}

\item{fact}{(default : NULL), Name of the factor in \code{physeq@sam_data}.
If set to NULL use the \code{left_name} and \code{right_name} parameter as modality.}

\item{merge_sample_by}{(default : NULL)}

\item{left_name}{(default : "A"): Name fo the left sample.}

\item{right_name}{(default : "B"): Name fo the right sample.}

\item{left_fill}{: Fill fo the left sample.}

\item{left_col}{: Color fo the left sample.}

\item{right_fill}{: Fill fo the right sample.}

\item{right_col}{: Color fo the right sample.}

\item{log_10}{(default : TRUE) : Does abundancy is log10 transformed ?}

\item{nudge_y}{: A parameter to control the y position of abundancy values.}

\item{geomLabel}{(default FALSE) : if true use the \code{geom_label} function
instead of \code{geom_text} to indicate the numbers of sequences}

\item{text_size}{(default : 3): default size for the number of sequences}

\item{...}{: other arguments for the ggplot function}
}
\value{
A plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\author{
Adrien Taudière
}
