% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{biplot_pq}
\alias{biplot_pq}
\title{Visualization of two samples for comparison}
\usage{
biplot_pq(
  physeq,
  fact = NULL,
  merge_sample_by = NULL,
  inverse_side = FALSE,
  left_name = NULL,
  right_name = NULL,
  left_fill = "#4B3E1E",
  left_col = "#f3f2d9",
  right_fill = "#1d2949",
  right_col = "#1d2949",
  log_10 = TRUE,
  nudge_y = 0.3,
  geomLabel = FALSE,
  text_size = 3,
  plotly_version = FALSE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{fact, }{Name of the factor in \code{physeq@sam_data}.
If left to NULL use the \code{left_name} and \code{right_name} parameter as modality.}

\item{merge_sample_by}{(default: NULL): if not \code{NULL} samples of
physeq are mereged using the vector set by \code{merge_sample_by}. This
merging used the \code{\link[speedyseq:merge_samples2]{speedyseq::merge_samples2()}}. In the case of
\code{\link[=biplot_pq]{biplot_pq()}} this must be a factor with two levels only.}

\item{inverse_side}{Inverse the side (put the right modality in the left side).}

\item{left_name}{Name fo the left sample.}

\item{right_name}{Name fo the right sample.}

\item{left_fill}{Fill fo the left sample.}

\item{left_col}{Color fo the left sample.}

\item{right_fill}{Fill fo the right sample.}

\item{right_col}{Color fo the right sample.}

\item{log_10}{(logical) Does abundancy is log10 transformed ?}

\item{nudge_y}{A parameter to control the y position of abundancy values.}

\item{geomLabel}{(default: FALSE, logical) if TRUE use the \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}} function
instead of \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} to indicate the numbers of sequences.}

\item{text_size}{default size for the number of sequences}

\item{plotly_version}{If TRUE, use \code{\link[plotly:ggplotly]{plotly::ggplotly()}} to return
a interactive ggplot.}

\item{...}{other arguments for the ggplot function}
}
\value{
A plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\author{
Adrien Taudière
}
