% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R, R/dada_phyloseq.R
\name{blast_pq}
\alias{blast_pq}
\title{Blast all sequence of \code{refseq} slot of a \code{\link{phyloseq-class}}
object against a custom database.}
\usage{
blast_pq(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 50,
  min_cover_cut = 50,
  unique_per_seq = FALSE,
  score_filter = TRUE,
  nproc = 1
)

blast_pq(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 50,
  min_cover_cut = 50,
  unique_per_seq = FALSE,
  score_filter = TRUE,
  nproc = 1
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{fasta_for_db}{path to a fasta file to make the blast database}

\item{database}{path to a blast database}

\item{id_cut}{(default: 90) cut of in identity percent to keep result}

\item{bit_score_cut}{(default: 50) cut of in bit score to keep result}

\item{min_cover_cut}{(default: 50) cut of in query cover (\%) to keep result}

\item{unique_per_seq}{(logical) if TRUE only return the first match for
each sequence in seq2search}

\item{score_filter}{(logical) does results are filter by score? If
FALSE, \code{id_cut},\code{bit_score_cut} and \code{min_cover_cut} are ignored}

\item{nproc}{(default: 1)
Set to number of cpus/processors to use for blast (args -num_threads
for blastn command)}
}
\value{
a blast table

a blast table
}
\description{
\verb{r lifecycle::badge("experimental")}

\verb{r lifecycle::badge("experimental")}
}
\seealso{
\code{\link[=blast_to_phyloseq]{blast_to_phyloseq()}} to use \code{refseq}
slot as a database

\code{\link[=blast_to_phyloseq]{blast_to_phyloseq()}} to use \code{refseq}
slot as a database
}
