% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_to_derep}
\alias{blast_to_derep}
\title{Blast some sequence against sequences from of a \code{\link{derep-class}}
object.}
\usage{
blast_to_derep(
  derep,
  seq2search,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 50,
  min_cover_cut = 50,
  unique_per_seq = FALSE,
  score_filter = FALSE,
  list_no_output_query = FALSE,
  min_length_seq = 200,
  args_makedb = NULL,
  args_blastn = NULL
)
}
\arguments{
\item{derep}{The result of \code{dada2::derepFastq()}. A list of \code{derep-class} object.}

\item{seq2search}{(required) path to a fasta file defining the sequences
you want to blast against the ASV sequences from the physeq object.}

\item{blastpath}{path to blast program}

\item{id_cut}{(default: 90) cut of in identity percent to keep result}

\item{bit_score_cut}{(default: 50) cut of in bit score to keep result}

\item{min_cover_cut}{(default: 50) cut of in query cover (\%) to keep result}

\item{unique_per_seq}{(logical) if TRUE only return the first match for
each sequence in seq2search}

\item{score_filter}{(logical; default FALSE) does results are filter by score? If
FALSE, \code{id_cut},\code{bit_score_cut} and \code{min_cover_cut} are ignored.}

\item{list_no_output_query}{(logical) does the result table include
query sequences for which \code{blastn} does not find any correspondence?}

\item{min_length_seq}{(default: 200) Removed sequences with less than
\code{min_length_seq} from derep before blast. Set to 0 to discard filtering
sequences by length.}

\item{args_makedb}{Additional parameters parse to makeblastdb command}

\item{args_blastn}{Additional parameters parse to blastn command}
}
\value{
A blast table
}
\description{
\verb{r lifecycle::badge("experimental")}
}
\seealso{
\code{\link[=blast_pq]{blast_pq()}} to use \code{refseq} slot as query sequences
against un custom database and \code{\link[=blast_to_phyloseq]{blast_to_phyloseq()}}  to use
\code{refseq} slot as a database
}
\author{
Adrien Taudière
}
