% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{blast_to_phyloseq}
\alias{blast_to_phyloseq}
\title{Blast some sequence against \code{refseq} slot of a \code{\link{phyloseq-class}}
object.}
\usage{
blast_to_phyloseq(
  physeq,
  seq2search,
  blastpath = NULL,
  id_cut = 90,
  bit_score_cut = 50,
  unique_per_seq = FALSE,
  score_filter = TRUE,
  list_no_output_query = FALSE
)
}
\arguments{
\item{seq2search}{(required) path to fasta file}

\item{blastpath}{path to blast program}

\item{id_cut}{(default: 90) cut of in identity percent to keep result}

\item{bit_score_cut}{(default: 1e-10) cut of in bit score to keep result}

\item{unique_per_seq}{(logical) if TRUE only return the first match for
each sequence in seq2search}

\item{score_filter}{(logical) does results are filter by score? If
FALSE, \code{id_cut} and \code{bit_score_cut} are ignored}

\item{list_no_output_query}{(logical) does the result table include
query sequences for which \code{blastn} does not find any correspondence?}
}
\value{
the blast table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
\dontrun{
data(data_fungi)
blastpath <- "...YOUR_PATH_TO_BLAST..."
blast_to_phyloseq(data_fungi,
  seq2search = "inst/extdata/ex.fasta",
  blastpath = blastpath
)
}
}
