% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{clean_physeq}
\alias{clean_physeq}
\title{Clean phyloseq object by removing empty samples and taxa
In addition, this function check for discrepancy (and rename) between
(i) taxa names in refseq, taxonomy table and otu_table and between
(ii) sample names in sam_data and otu_table.}
\usage{
clean_physeq(
  physeq,
  remove_empty_samples = TRUE,
  remove_empty_taxa = TRUE,
  clean_samples_names = TRUE
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{dada2} package}

\item{remove_empty_samples}{(logical, default TRUE) : Do you want to remove samples without sequences (this is done after removing empty taxa)}

\item{remove_empty_taxa}{(logical, default TRUE) : Do you want to remove taxa without sequences (this is done before removing empty samples)}

\item{clean_samples_names}{: Do you want to clean samples names?}
}
\value{
A new \code{\link{phyloseq-class}} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
