% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{compare_pairs_pq}
\alias{compare_pairs_pq}
\title{Compare samples in pairs using diversity and number of ASV including
shared ASV.}
\usage{
compare_pairs_pq(
  physeq = NULL,
  modality = NULL,
  bifactor = NULL,
  nb_min_seq = 0,
  vegIndex = "shannon"
)
}
\arguments{
\item{physeq}{(required) A \code{\link{phyloseq-class}} object.
For the moment refseq slot need to be not Null.}

\item{modality}{the name of the column in the \code{sam_data}
slot of the physeq object to split samples by pairs}

\item{bifactor}{(required) a factor (present in the \code{sam_data} slot of
the physeq object) presenting the pair names}

\item{nb_min_seq}{minimum number of sequences per sample
to count the ASV/OTU}

\item{vegIndex}{(default: "shannon") index for the \code{vegan::diversity} function}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
