% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_misc.R
\name{cutadapt_remove_primers}
\alias{cutadapt_remove_primers}
\title{Remove primers using cutapapt Work In Progress}
\usage{
cutadapt_remove_primers(
  path,
  output_folder = "wo_primers",
  primer_fw = NULL,
  primer_rev = NULL,
  nproc = 1,
  ...
)
}
\value{
a list of commande but in The future, will
run the commandes using system
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Adrien Taudière
}
