% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fungi}
\alias{data_fungi}
\title{Some fungal OTU in phyloseq format}
\format{
A physeq object containing 1420 taxa with references sequences
described by 14 taxonomic ranks
and 185 samples described by 14 sample variables:
}
\usage{
data_fungi
}
\description{
A dataset
}
\keyword{datasets}
