% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fungi_sp_known}
\alias{data_fungi_sp_known}
\title{Some fungal OTU in phyloseq format with known species name}
\format{
A physeq object containing 651 taxa with references sequences
described by 14 taxonomic ranks
and 185 samples described by 14 sample variables:
}
\usage{
data_fungi_sp_known
}
\description{
A subset of the data_fungi dataset including only taxa with information
at the species level
}
\keyword{datasets}
