% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{dist_pos_control}
\alias{dist_pos_control}
\title{Calculate ecological distance among positive controls vs
  distance for all samples}
\usage{
dist_pos_control(physeq, samples_names, method = "bray")
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{samples_names}{(required): a vector of names for samples with
positives controls of the same samples having the same name}

\item{method}{(default = "bray"): a method to calculate
the distance, parsed to vegdist}
}
\value{
A list of two dataframes with
  (i) the distance among positive controls and
  (ii) the distance among all samples
}
\description{
#' `r lifecycle::badge("maturing")`
}
\details{
Compute distance among positive controls,
  i.e. samples which are duplicated
  to test for variation, for example in
   (i) a step in the sampling,
   (ii) a step in the extraction,
   (iii) a step in the sequencing.
}
\author{
Adrien Taudière
}
