% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R, R/dada_phyloseq.R
\name{filter_asv_blast}
\alias{filter_asv_blast}
\title{Filter indesirable taxa using blast against a against a custom database.}
\usage{
filter_asv_blast(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  clean_pq = TRUE,
  blastpath = NULL,
  id_cut = 80,
  bit_score_cut = 150,
  min_cover_cut = 50,
  add_info_to_taxtable = TRUE,
  nproc = 1
)

filter_asv_blast(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  clean_pq = TRUE,
  blastpath = NULL,
  id_cut = 80,
  bit_score_cut = 150,
  min_cover_cut = 50,
  add_info_to_taxtable = TRUE,
  nproc = 1
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{fasta_for_db}{path to a fasta file to make the blast database}

\item{database}{path to a blast database}

\item{clean_pq}{(logical)
If set to TRUE, empty samples and empty ASV are discarded
after filtering.}

\item{blastpath}{path to blast program}

\item{id_cut}{(default: 80) cut of in identity percent to keep ASV}

\item{bit_score_cut}{(default: 150) cut of in bit score to keep result}

\item{min_cover_cut}{(default: 50) cut of in query cover (\%) to keep result}

\item{nproc}{(default: 1)
Set to number of cpus/processors to use for blast (args -num_threads
for blastn command)}

\item{...}{Others options for the \code{balst_pq()} function. See \code{?blast_pq}.
Note that params \code{unique_per_seq} and \code{score_filter} must be set to TRUE.}

\item{add_info_to_taxtable:}{add some info from blast query to taxtable of the
new physeq object. Only the information ("Query name", "Taxa name", "bit score",
"\% id. match", "Query cover", "e-value") for higher e-value hit.
for each ASV is add to taxtable. Note that query name may be different from
final taxa names as some function proposed to change the ASV names.}
}
\value{
A new \code{\link{phyloseq-class}} object.

A new \code{\link{phyloseq-class}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
