% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{funky_color}
\alias{funky_color}
\title{Count sequences in fasta or fastq file}
\usage{
funky_color(n)
}
\arguments{
\item{n}{a number of colors}
}
\value{
a color palette
}
\description{
Count sequences in fasta or fastq file
}
\author{
Thibaut Jombart
}
