% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(file_path)
}
\arguments{
\item{file_path}{(required): path to a file}
}
\value{
The extension of a file.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}


Internally used in \code{\link[=count_seq]{count_seq()}}.
Warning: don't work when there is '.' in the name of the
file before the extension
}
\author{
Adrien Taudière
}
