% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(file)
}
\arguments{
\item{file}{(required): path to a file}
}
\value{
A  \code{\link{phyloseq-class}} object with simplified taxonomy
}
\description{
\verb{r lifecycle::badge("maturing")}
}
\author{
Adrien Taudière
}
