% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(file_path)
}
\arguments{
\item{file_path}{(required): path to a file}
}
\value{
A  \code{\link{phyloseq-class}} object with simplified taxonomy
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}
}
\author{
Adrien Taudière
}
