% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggvenn_pq}
\alias{ggvenn_pq}
\title{Venn diagram of \code{\link{phyloseq-class}} object using
\code{ggVennDiagram::ggVennDiagram} function}
\usage{
ggvenn_pq(
  physeq = NULL,
  fact = NULL,
  merge_sample_by = NULL,
  min_nb_seq = 0,
  taxonomic_rank = NULL,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{fact}{(required): Name of the factor to cluster samples by modalities.
Need to be in \code{physeq@sam_data}.}

\item{merge_sample_by}{(default: NULL): if not \code{NULL} samples of
physeq are mereged using the vector set by \code{merge_sample_by}. This
merging used the \code{\link[speedyseq:merge_samples2]{speedyseq::merge_samples2()}}.
If \code{merge_sample_by} is equal to \code{fact}
it give the same result as if \code{merge_sample_by} is NULL.}

\item{min_nb_seq}{(default: 0)): minimum number of sequences by OTUs by
samples to take into count this OTUs in this sample. For example,
if min_nb_seq=2,each value of 2 or less in the OTU table
will not count in the venn diagramm}

\item{taxonomic_rank}{(default: Null): Name (or number) of a taxonomic rank
to count. If set to Null (the default) the number of OTUs is counted.}

\item{...}{other arguments for the \code{ggVennDiagram::ggVennDiagram} function
for ex. \code{category.names}.}
}
\value{
A \code{\link{ggplot}}2 plot representing Venn diagramm of
modalities of the argument \code{factor}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Note that you can use ggplot2 function to customize the plot
for ex. \code{+ scale_fill_distiller(palette = "BuPu", direction = 1)}
}
\author{
Adrien Taudière
}
