% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggvenn_pq}
\alias{ggvenn_pq}
\title{Venn diagram of \code{\link{phyloseq-class}} object using
\code{ggVennDiagram::ggVennDiagram} function}
\usage{
ggvenn_pq(
  physeq = NULL,
  fact = NULL,
  min_nb_seq = 0,
  taxonomic_rank = NULL,
  split_by = NULL,
  add_nb_samples = TRUE,
  add_nb_seq = FALSE,
  rarefy_before_merging = FALSE,
  rarefy_after_merging = FALSE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{fact}{(required): Name of the factor to cluster samples by modalities.
Need to be in \code{physeq@sam_data}.}

\item{min_nb_seq}{minimum number of sequences by OTUs by
samples to take into count this OTUs in this sample. For example,
if min_nb_seq=2,each value of 2 or less in the OTU table
will not count in the venn diagram}

\item{taxonomic_rank}{Name (or number) of a taxonomic rank
to count. If set to Null (the default) the number of OTUs is counted.}

\item{split_by}{Split into multiple plot using variable split_by.
The name of a variable must be present in \code{sam_data} slot
of the physeq object.}

\item{add_nb_samples}{(logical, default TRUE) Add the number of samples to
levels names}

\item{add_nb_seq}{(logical, default FALSE) Add the number of sequences to
levels names}

\item{rarefy_before_merging}{Rarefy each sample before merging by the
modalities of args \code{fact}. Use \code{phyloseq::rarefy_even_depth()} function}

\item{rarefy_after_merging}{Rarefy each sample after merging by the
modalities of args \code{fact}.}

\item{...}{Other arguments for the \code{ggVennDiagram::ggVennDiagram} function
for ex. \code{category.names}.}
}
\value{
A \code{\link{ggplot}}2 plot representing Venn diagram of
modalities of the argument \code{factor} or if split_by is set a list
of plots.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Note that you can use ggplot2 function to customize the plot
for ex. \code{+ scale_fill_distiller(palette = "BuPu", direction = 1)}
and \code{+ scale_x_continuous(expand = expansion(mult = 0.5))}. See
examples.
}
\examples{
if (requireNamespace("ggVennDiagram")) {
  ggvenn_pq(data_fungi, fact = "Height")
}
\donttest{
if (requireNamespace("ggVennDiagram")) {
  ggvenn_pq(data_fungi, fact = "Height") +
    ggplot2::scale_fill_distiller(palette = "BuPu", direction = 1)
  pl <- ggvenn_pq(data_fungi, fact = "Height", split_by = "Time")
  for (i in 1:length(pl)) {
    p <- pl[[i]] +
      scale_fill_distiller(palette = "BuPu", direction = 1) +
      theme(plot.title = element_text(hjust = 0.5, size = 22))
    print(p)
  }

  data_fungi2 <- subset_samples(data_fungi, data_fungi@sam_data$Tree_name == "A10-005" |
    data_fungi@sam_data$Height \%in\% c("Low", "High"))
  ggvenn_pq(data_fungi2, fact = "Height")

  ggvenn_pq(data_fungi, fact = "Height", add_nb_seq = TRUE, set_size = 4)
  ggvenn_pq(data_fungi, fact = "Height", rarefy_before_merging = TRUE)
  ggvenn_pq(data_fungi, fact = "Height", rarefy_after_merging = TRUE) +
    scale_x_continuous(expand = expansion(mult = 0.5))
}
}
}
\seealso{
\code{\link[=upset_pq]{upset_pq()}}
}
\author{
Adrien Taudière
}
