% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{hill_pq}
\alias{hill_pq}
\title{Graphical representation of hill number 0, 1 and 2 across a factor}
\usage{
hill_pq(physeq, variable, color_fac = NA, letters = FALSE)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}} object}

\item{variable}{(required): The variable to test}

\item{color_fac}{(optional): The variable to color the barplot}

\item{letters}{(optional, default=FALSE): If set to TRUE, the plot
show letters based on p-values for comparison. Use the
\code{\link[multcompView]{multcompLetters}} function from the package
multcompLetters.}
}
\value{
A list of 4 ggplot2 plot.
\itemize{
\item plot_Hill_0 : the boxplot of Hill number 0 (= species richness)
against the variable
\item plot_Hill_1 : the boxplot of Hill number 1 (= Shannon index)
against the variable
\item plot_Hill_2 : the boxplot of Hill number 2 (= Simpson index)
against the variable
\item plot_tuckey : plot the result of the Tuckey HSD test
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
