% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{iNEXT_pq}
\alias{iNEXT_pq}
\title{iNterpolation and EXTrapolation of Hill numbers (with iNEXT)}
\usage{
iNEXT_pq(physeq, merge_sample_by = NULL, ...)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{merge_sample_by}{(default: NULL) if not \code{NULL} samples of
physeq are merged using the vector set by \code{merge_sample_by}. This
merging used the \code{\link[speedyseq:merge_samples2]{speedyseq::merge_samples2()}}. In the case of
\code{\link[=biplot_pq]{biplot_pq()}} this must be a factor with two levels only.}

\item{...}{other arguments for the \code{\link[iNEXT:iNEXT]{iNEXT::iNEXT()}} function}
}
\value{
see \code{\link[iNEXT:iNEXT]{iNEXT::iNEXT()}} documentation
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
library("iNEXT")
res_iNEXT <- iNEXT_pq(
  data_fungi_sp_known,
  merge_sample_by = "Height",
  q = 1,
  datatype = "abundance",
  nboot = 2
)
ggiNEXT(res_iNEXT)
ggiNEXT(res_iNEXT, type = 2)
ggiNEXT(res_iNEXT, type = 3)
}
}
\author{
Adrien Taudière
}
