% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_cutadapt_installed}
\alias{is_cutadapt_installed}
\title{Test if cutadapt is installed.}
\usage{
is_cutadapt_installed(
  args_before_cutadapt =
    "source ~/miniconda3/etc/profile.d/conda.sh && conda activate cutadaptenv && "
)
}
\arguments{
\item{args_before_cutadapt}{: (String) A one line bash command to run before
to run cutadapt. For examples, "source ~/miniconda3/etc/profile.d/conda.sh && conda activate cutadaptenv &&" allow to bypass the conda init which asks to restart the shell}
}
\value{
A logical that say if cutadapt is install in
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_cutadapt_installed()
}
\author{
Adrien Taudière
}
