% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_krona_installed}
\alias{is_krona_installed}
\title{Test if krona is installed.}
\usage{
is_krona_installed(path = "ktImportKrona")
}
\arguments{
\item{path}{(default: krona) Path to krona}
}
\value{
A logical that say if krona is install in
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_krona_installed()
}
\author{
Adrien Taudière
}
