% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{is_vsearch_installed}
\alias{is_vsearch_installed}
\title{Test if vsearch is installed.}
\usage{
is_vsearch_installed(path = "vsearch")
}
\arguments{
\item{path}{(default: vsearch) Path to vsearch}
}
\value{
A logical that say if vsearch is install in
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturing"></a>
}}


Useful for testthat and examples compilation for R CMD CHECK and
test coverage
}
\examples{
MiscMetabar::is_vsearch_installed()
}
\author{
Adrien Taudière
}
