% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{merge_krona}
\alias{merge_krona}
\title{Merge Krona files using \href{https://github.com/marbl/Krona/wiki}{KronaTools}.}
\usage{
merge_krona(files = NULL, output = "mergeKrona.html")
}
\arguments{
\item{files}{(required) path to html files to merged}

\item{output}{path to the output file}
}
\value{
A html file
}
\description{
\verb{r lifecycle::badge("maturing")}

Need the installation of kronatools on the computer (\href{https://github.com/marbl/Krona/wiki/Installing}{installation instruction}).
Function merge_krona allows merging multiple html files in one interactive krona file
}
\examples{
data("GlobalPatterns")
GA <- subset_taxa(GlobalPatterns, Phylum == "Acidobacteria")
# krona(GA, "Number.of.sequences.html")
# krona(GA, "Number.of.ASVs.html", nb_seq = F)
# merge_krona(c("Number.of.sequences.html", "Number.of.ASVs.html"))
}
\seealso{
\code{\link{krona}}
}
\author{
Adrien Taudière
}
