% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{MiscMetabar-deprecated}
\alias{MiscMetabar-deprecated}
\alias{physeq_graph_test}
\alias{adonis_phyloseq}
\alias{clean_physeq}
\alias{lulu_phyloseq}
\alias{otu_circle}
\alias{biplot_physeq}
\alias{read_phyloseq}
\alias{write_phyloseq}
\alias{sankey_phyloseq}
\alias{summary_plot_phyloseq}
\alias{plot_deseq2_phyloseq}
\alias{plot_edgeR_phyloseq}
\alias{venn_phyloseq}
\alias{ggVenn_phyloseq}
\alias{hill_tuckey_phyloseq}
\alias{hill_phyloseq}
\title{Deprecated function(s) in the MiscMetabar package}
\usage{
physeq_graph_test(...)
}
\arguments{
\item{...}{Parameters to be passed on to the modern version of the function}
}
\value{
Depend on the functions.
}
\description{
These functions are provided for compatibility with older version of
the MiscMetabar package. They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
\link{graph_test_pq} \tab now a synonym for \code{physeq_graph_test}\cr
\link{adonis_pq} \tab now a synonym for \code{adonis_phyloseq}\cr
\link{clean_pq} \tab now a synonym for \code{clean_physeq}\cr
\link{lulu_pq} \tab now a synonym for \code{lulu_phyloseq}\cr
\link{circle_pq} \tab now a synonym for \code{otu_circle}\cr
\link{biplot_pq} \tab now a synonym for \code{biplot_physeq}\cr
\link{read_pq} \tab now a synonym for \code{read_phyloseq}\cr
\link{write_pq} \tab now a synonym for \code{write_phyloseq}\cr
\link{sankey_pq} \tab now a synonym for \code{sankey_phyloseq}\cr
\link{summary_plot_pq} \tab now a synonym for \code{summary_plot_phyloseq}\cr
\link{plot_edgeR_pq} \tab now a synonym for \code{plot_edgeR_phyloseq}\cr
\link{plot_deseq2_pq} \tab now a synonym for \code{plot_deseq2_phyloseq}\cr
\link{venn_pq} \tab now a synonym for \code{venn_phyloseq}\cr
\link{ggvenn_pq} \tab now a synonym for \code{ggVenn_phyloseq}\cr
\link{hill_tuckey_pq} \tab now a synonym for \code{hill_tuckey_phyloseq}\cr
\link{hill_pq} \tab now a synonym for \code{hill_phyloseq}\cr
heat_tree_pq \tab now a synonym for \code{physeq_heat_tree}\cr
\link{compare_pairs_pq} \tab now a synonym for \code{multiple_share_bisamples}\cr
}
}

