% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{MiscMetabar-deprecated}
\alias{MiscMetabar-deprecated}
\alias{physeq_graph_test}
\alias{adonis_phyloseq}
\alias{clean_physeq}
\alias{lulu_phyloseq}
\alias{otu_circle}
\alias{biplot_physeq}
\alias{read_phyloseq}
\alias{write_phyloseq}
\title{Deprecated function(s) in the MiscMetabar package}
\usage{
physeq_graph_test(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the MiscMetabar package. They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
\link{graph_test_pq} \tab now a synonym for \link{physeq_graph_test}\cr
\link{adonis_pq} \tab now a synonym for \link{adonis_phyloseq}\cr
\link{clean_pq} \tab now a synonym for \link{clean_physeq}\cr
\link{lulu_pq} \tab now a synonym for \link{lulu_phyloseq}\cr
\link{circle_pq} \tab now a synonym for \link{otu_circle}\cr
\link{biplot_pq} \tab now a synonym for \link{biplot_physeq}\cr
\link{read_pq} \tab now a synonym for \link{read_phyloseq}\cr
\link{write_pq} \tab now a synonym for \link{write_phyloseq}\cr
}
}

