% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscMetabar-package.R
\name{MiscMetabar-package}
\alias{MiscMetabar-package}
\title{\code{MiscMetabar} package}
\description{
Functions to help analyze and visualize metabarcoding data. Mainly based on
the phyloseq and dada2 packages.
}
