% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscMetabar-package.R
\docType{package}
\name{MiscMetabar-package}
\alias{MiscMetabar}
\alias{MiscMetabar-package}
\title{MiscMetabar: Miscellaneous functions for metabarcoding analysis}
\description{
Functions to help analyse and visualise metabarcoding data. Mainly based on the phyloseq and dada2 packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adrientaudiere/MiscMetabar}
  \item Report bugs at \url{https://github.com/adrientaudiere/MiscMetabar/issues}
}

}
\keyword{internal}
