% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{: list of ggplot objects}

\item{plotlist}{: list of ggplot objects}

\item{cols}{: number of columns}

\item{layout}{: A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
`r lifecycle::badge("stable")`
}
