% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ordination_pq}
\alias{plot_ordination_pq}
\title{A wrapper of plot_ordination with vegan distance matrix}
\usage{
plot_ordination_pq(
  physeq,
  method = "robust.aitchison",
  ordination_method = "NMDS",
  ...
)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{method}{(string, default "robust.aitchison") The distance method to use
from vegan::vegdist(). See ?vegan::vegdist for more details.}

\item{ordination_method}{(string, default "NMDS") The ordination method to use
in phyloseq::ordinate(). See ?phyloseq::ordinate for more details.}

\item{...}{Additional arguments passed on to phyloseq::plot_ordination()}
}
\value{
A ggplot2 object
}
\description{
A wrapper of plot_ordination with vegan distance matrix
}
\details{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Basically a wrapper of \code{\link[phyloseq:plot_ordination]{phyloseq::plot_ordination()}} to use aitchison and
robust.aitchison distances from vegan package.
}
\examples{
library(patchwork)
plot_ordination_pq(data_fungi_mini, method = "robust.aitchison", color = "Height") +
  plot_ordination_pq(data_fungi_mini, method = "bray", color = "Height")
}
\author{
Adrien Taudière
}
