% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_misc.R
\name{rename_samples_otu_table}
\alias{rename_samples_otu_table}
\title{Rename samples of an otu_table}
\usage{
rename_samples_otu_table(otu_tab, names_of_samples, taxa_are_rows = FALSE)
}
\arguments{
\item{otu_tab}{(required) The matrix or otu_table}

\item{names_of_samples}{(required) The}

\item{taxa_are_rows}{(default : FALSE) Does the taxa are rows or
columns.}
}
\value{
the matrix with new colnames
(or rownames if \code{taxa_are_rows} is true)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data(data_fungi)
rename_samples_otu_table(data_fungi@otu_table, as.character(1:185),
  taxa_are_rows = T
)
}
