% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reshape_ggplot}
\alias{reshape_ggplot}
\title{Reshape a ggplot2 object by wrapping its text elements}
\usage{
reshape_ggplot(
  plot,
  width = 60,
  width_subtitle = width * fontsize_title/fontsize_subtitle,
  width_labs = width * fontsize_title/fontsize_labs,
  width_caption = width * fontsize_title/fontsize_caption,
  fontsize_title = 10,
  fontsize_subtitle = 7,
  fontsize_labs = 8,
  fontsize_caption = 7
)
}
\arguments{
\item{plot}{(required) A ggplot2 object.}

\item{width}{(int, default 60) The wrapping width (in characters) of the
title. Also used to compute the default value of the other widths.}

\item{width_subtitle}{(int, default \code{width * fontsize_title / fontsize_subtitle}) The wrapping width of the subtitle.}

\item{width_labs}{(int, default \code{width * fontsize_title / fontsize_labs})
The wrapping width of the x and y axis labels.}

\item{width_caption}{(int, default \code{width * fontsize_title / fontsize_caption}) The wrapping width of the caption.}

\item{fontsize_title}{(int, default 10) Font size for the title.}

\item{fontsize_subtitle}{(int, default 7) Font size for the subtitle.}

\item{fontsize_labs}{(int, default 8) Font size for the x and y axis labels.}

\item{fontsize_caption}{(int, default 7) Font size for the caption.}
}
\value{
A ggplot2 object with wrapped/resized text elements.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Auto-wrap (and optionally resize the font of) the title, subtitle, caption
and axis labels of a ggplot2 plot using \code{\link[stringr:str_wrap]{stringr::str_wrap()}}. Useful to tidy
up long titles produced automatically by other \code{MiscMetabar} plotting
functions, in particular before combining several plots with \code{patchwork}.
}
\examples{
\donttest{
if (requireNamespace("stringr") && requireNamespace("patchwork")) {
  df_mini <- prune_samples(
    sample_names(data_fungi_mini)[1:5],
    data_fungi_mini
  )
  patchwork::wrap_plots(lapply(
    hill_pq(df_mini, "Height", vioplot = TRUE),
    reshape_ggplot,
    width = 45
  ))
}
}
}
\author{
Adrien Taudière
}
