% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ridges_pq}
\alias{ridges_pq}
\title{Ridge plot of a phyloseq object}
\usage{
ridges_pq(physeq, fact, nb_seq = TRUE, log10trans = TRUE, ...)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{fact}{(required) Name of the factor in \code{physeq@sam_data} used to plot
different lines}

\item{nb_seq}{(logical; default TRUE) If set to FALSE, only the number of ASV
is count. Concretely, physeq \code{otu_table} is transformed in a binary
\code{otu_table} (each value different from zero is set to one)}

\item{log10trans}{(logical, default TRUE) If TRUE,
the number of sequences (or ASV if nb_seq = FALSE) is log10
transformed.}

\item{...}{Other params passed on to \code{\link[ggridges:geom_density_ridges]{ggridges::geom_density_ridges()}}}
}
\value{
A \code{\link{ggplot}}2 plot  with bar representing the number of sequence en each
taxonomic groups
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# data(data_fungi)
ridges_pq(data_fungi, "Time", alpha = 0.5, log10trans = FALSE) + xlim(c(0, 1000))
ridges_pq(data_fungi, "Time", alpha = 0.5)
ridges_pq(
  clean_pq(subset_taxa(data_fungi_sp_known, Phylum == "Basidiomycota")),
  "Sample_names"
)
ridges_pq(clean_pq(subset_taxa(data_fungi_sp_known, Phylum == "Basidiomycota")),
  "Time",
  alpha = 0.6, scale = 0.9
)
ridges_pq(clean_pq(subset_taxa(data_fungi_sp_known, Phylum == "Basidiomycota")),
  "Time",
  jittered_points = TRUE,
  position = ggridges::position_points_jitter(width = 0.05, height = 0),
  point_shape = "|", point_size = 3, point_alpha = 1, alpha = 0.7,
  scale = 0.8
)
ridges_pq(data_fungi, "Height", alpha = 0.5, log10trans = TRUE)
}
\author{
Adrien Taudière
}
