% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{simplify_taxo}
\alias{simplify_taxo}
\title{Simplify taxonomy by removing some unused characters such as "k__"}
\usage{
simplify_taxo(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
A  \code{\link{phyloseq-class}} object with simplified taxonomy
}
\description{
\verb{r lifecycle::badge("maturing")}
}
\author{
Adrien Taudière
}
