% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{simplify_taxo}
\alias{simplify_taxo}
\title{Simplify taxonomy by removing some unused character such as "k__"}
\usage{
simplify_taxo(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}
}
\value{
A  \code{\link{phyloseq-class}} object with simplified taxonomy
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\author{
Adrien Taudière
}
