% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{subsample_fastq}
\alias{subsample_fastq}
\title{Subsample a fastq file copying the n_seq first sequences in a given folder}
\usage{
subsample_fastq(fastq_files, folder_output = "subsample", n_seq = 1000)
}
\arguments{
\item{folder_output}{The path to a folder for output files}

\item{n_seq}{(int; default 1000) : Number of sequences kept (every sequence
spread across 4 lines)}

\item{file_path}{The path to one fastq file or a list of fastq files
(see examples)}
}
\value{
Nothing, create subsampled fastq files in a folder
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
subsample_fastq("inst/extdata/ex_R1_001.fastq.gz", "your_path_to_output")
subsample_fastq(list_fastq_files("inst/extdata"), "your_path_to_output", n = 10)
}
}
\author{
Adrien Taudière
}
