% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellanous.R
\name{subsample_fastq}
\alias{subsample_fastq}
\title{Subsample a fastq file copying the n_seq first sequences in a given folder}
\usage{
subsample_fastq(fastq_files, folder_output = "subsample", nb_seq = 1000)
}
\arguments{
\item{fastq_files}{The path to one fastq file or a list of fastq files
(see examples)}

\item{folder_output}{The path to a folder for output files}

\item{nb_seq}{(int; default 1000) : Number of sequences kept (every sequence
spread across 4 lines)}
}
\value{
Nothing, create subsampled fastq files in a folder
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
ex_file <- system.file("extdata", "ex_R1_001.fastq.gz",
  package = "MiscMetabar",
  mustWork = TRUE
)
subsample_fastq(ex_file, "your_path_to_output")
subsample_fastq(list_fastq_files("extdata"), "your_path_to_output", n = 10)
}
}
\author{
Adrien Taudière
}
