% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary_plot_phyloseq}
\alias{summary_plot_phyloseq}
\title{Summarise a \code{\link{phyloseq-class}} object using a plot.}
\usage{
summary_plot_phyloseq(physeq)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}} object}
}
\value{
A ggplot2 object
}
\description{
Summarise a \code{\link{phyloseq-class}} object using a plot.
}
