% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{summary_plot_phyloseq}
\alias{summary_plot_phyloseq}
\title{Summarise a \code{\link{phyloseq-class}} object using a plot.}
\usage{
summary_plot_phyloseq(physeq)
}
\arguments{
\item{physeq}{(required): A \code{\link{phyloseq-class}} object.
For the moment refseq slot need to be not Null.}
}
\value{
A ggplot2 object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
data(data_fungi)
summary_plot_phyloseq(data_fungi)
}
