% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{taxa_as_rows}
\alias{taxa_as_rows}
\title{Force taxa to be in columns in the otu_table of a physeq object}
\usage{
taxa_as_rows(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
A new \code{\link{phyloseq-class}} object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-maturing-blue" alt="lifecycle-maturingalt="lifecycle-maturing"></a>
}}
}
\author{
Adrien Taudière
}
