% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{track_wkflow}
\alias{track_wkflow}
\title{Track the number of reads (= sequences), samples and cluster (e.g. ASV)
from various objects including dada-class and derep-class.}
\usage{
track_wkflow(list_of_objects, obj_names = NULL)
}
\arguments{
\item{list_of_objects}{(required): A list of objects}

\item{obj_names}{(default: NULL) :
A list of names corresponding to the list of objects}
}
\value{
The number of sequences, clusters (e.g. OTUs, ASVs) and samples for
  each object.
}
\description{
- List of fastq and fastg.gz files -> nb of reads and samples
 - List of dada-class -> nb of reads, clusters (ASV) and samples
 - List of derep-class -> nb of reads, clusters (unique sequences) and samples
 - Matrix of samples x clusters (e.g. otu_table) -> nb of reads, clusters and samples
 - Phyloseq-class -> nb of reads, clusters and samples
}
