% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{track_wkflow}
\alias{track_wkflow}
\title{Track the number of reads (= sequences), samples and cluster (e.g. ASV)
from various objects including dada-class and derep-class.}
\usage{
track_wkflow(list_of_objects, obj_names = NULL, clean_pq = FALSE)
}
\arguments{
\item{list_of_objects}{(required) a list of objects}

\item{obj_names}{A list of names corresponding to the list of objects}

\item{clean_pq}{(logical)
If set to TRUE, empty samples and empty ASV are discarded
before clustering.}
}
\value{
The number of sequences, clusters (e.g. OTUs, ASVs) and samples for
each object.
}
\description{
Track the number of reads (= sequences), samples and cluster (e.g. ASV)
from various objects including dada-class and derep-class.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
\itemize{
\item List of fastq and fastg.gz files -> nb of reads and samples
\item List of dada-class -> nb of reads, clusters (ASV) and samples
\item List of derep-class -> nb of reads, clusters (unique sequences)
and samples
\item Matrix of samples x clusters (e.g. \code{otu_table}) -> nb of reads,
clusters and samples
\item Phyloseq-class -> nb of reads, clusters and samples
}
}
