% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{track_wkflow}
\alias{track_wkflow}
\title{Track the number of reads (= sequences), samples and cluster (e.g. ASV)
from various objects including dada-class and derep-class.}
\usage{
track_wkflow(
  list_of_objects,
  obj_names = NULL,
  clean_pq = FALSE,
  taxonomy_rank = NULL,
  ...
)
}
\arguments{
\item{list_of_objects}{(required) a list of objects}

\item{obj_names}{A list of names corresponding to the list of objects}

\item{clean_pq}{(logical)
If set to TRUE, empty samples and empty ASV are discarded
before clustering.}

\item{taxonomy_rank}{A vector of int. Define the column number of
taxonomic rank \verb{in physeq@tax_table} to compute the number of unique value.
Default is NULL and do not compute values for any taxonomic rank}

\item{...}{Other arguments passed on to \code{\link[=clean_pq]{clean_pq()}} function.}
}
\value{
The number of sequences, clusters (e.g. OTUs, ASVs) and samples for
each object.
}
\description{
Track the number of reads (= sequences), samples and cluster (e.g. ASV)
from various objects including dada-class and derep-class.
}
\details{
\verb{r lifecycle::badge("maturing")}
\itemize{
\item List of fastq and fastg.gz files -> nb of reads and samples
\item List of dada-class -> nb of reads, clusters (ASV) and samples
\item List of derep-class -> nb of reads, clusters (unique sequences)
and samples
\item Matrix of samples x clusters (e.g. \code{otu_table}) -> nb of reads,
clusters and samples
\item Phyloseq-class -> nb of reads, clusters and samples
}
}
