% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsne_pq}
\alias{tsne_pq}
\title{Compute tSNE position of samples from a phyloseq object}
\usage{
tsne_pq(physeq, method = "bray", dims = 2, theta = 0, perplexity = 30, ...)
}
\arguments{
\item{physeq}{(required) a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{method}{A method to calculate distance using \code{vegan::vegdist()} function}

\item{dims}{(Int) Output dimensionality (default: 2)}

\item{theta}{(Numeric) Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.0 see details in the man page of \code{Rtsne::Rtsne}).}

\item{perplexity}{(Numeric) Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details in the man page of \code{Rtsne::Rtsne})}

\item{...}{Additional arguments passed on to \code{Rtsne::Rtsne()}}
}
\value{
A dataframe with samples informations and the x_tsne and y_tsne
position (or \code{tsne_1}, \code{tsne_2}, ... columns when \code{dims != 2}).
}
\description{
Compute tSNE position of samples from a phyloseq object
}
\examples{
\dontshow{if (tolower(Sys.info()[["sysname"]]) != "windows") withAutoprint(\{ # examplesIf}
if (requireNamespace("Rtsne")) {
  df_tsne <- tsne_pq(data_fungi_mini)
  ggplot(df_tsne, aes(x = x_tsne, y = y_tsne, col = Height)) +
    geom_point(size = 2)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}, \code{\link[=umap_pq]{umap_pq()}}, \code{\link[phyloseq:plot_ordination]{phyloseq::plot_ordination()}}
}
\author{
Adrien Taudiere
}
