% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{verify_pq}
\alias{verify_pq}
\title{Verify the validity of a phyloseq object}
\usage{
verify_pq(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
Nothing if the phyloseq object is valid. An error in the other case.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Mostly for internal use in MiscMetabar functions.
}
