% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{verify_pq}
\alias{verify_pq}
\title{Verify the validity of a phyloseq object}
\usage{
verify_pq(physeq)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
using the \code{phyloseq} package.}
}
\value{
Nothing if the phyloseq object is valid. An error in the other case.
}
\description{
Verify the validity of a phyloseq object
}
\details{
\verb{r lifecycle::badge("maturing")}

Mostly for internal use in MiscMetabar functions.
}
