% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{vsearch_search_global}
\alias{vsearch_search_global}
\title{Search for a list of sequence in a fasta file against physeq reference
  sequences using Vsearch}
\usage{
vsearch_search_global(
  physeq,
  seq2search,
  vsearchpath = "vsearch",
  id = 0.8,
  iddef = 0
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{seq2search}{(required): path to fasta file}

\item{vsearchpath}{(default = "vsearch"): path to vsearch}

\item{id}{(default = 0.8): id for --usearch_global}

\item{iddef}{(default = 0): iddef for --usearch_global}
}
\value{
A dataframe with uc results (invisible)
}
\description{
Search for a list of sequence in a fasta file against physeq reference
  sequences using Vsearch
}
