% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Save phyloseq object in the form of multiple csv tables.}
\usage{
write_phyloseq(physeq, path = NULL)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{path}{(defaut: NULL) : a path to the folder to save the phyloseq object}
}
\value{
Four csv tables (refseq.csv, otu_table.csv, tax_table.csv, sam_data.csv)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
