% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Save phyloseq object in the form of multiple csv tables. EXPERIMENTAL}
\usage{
write_phyloseq(physeq, path = NULL)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{path}{(defaut: NULL) : a path to the folder to save the phyloseq object}
}
\value{

}
\description{
Save phyloseq object in the form of multiple csv tables. EXPERIMENTAL
}
