% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_phyloseq.R
\name{write_phyloseq}
\alias{write_phyloseq}
\title{Save phyloseq object in the form of multiple csv tables.}
\usage{
write_phyloseq(
  physeq,
  path = NULL,
  rdata = FALSE,
  one_file_ASV = FALSE,
  write_sam_data = TRUE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link{phyloseq-class}} object.}

\item{rdata}{(logical): does the phyloseq object is also saved in Rdata format?}

\item{one_file_ASV}{(logical): if TRUE, combine all data in one file only}

\item{write_sam_data}{(logical): does the samples data are add to
the file. Only used if \code{one_file_ASV} is TRUE.
Note that this result in a lot of NA values.}

\item{...}{Other arguments passed to \code{\link[utils:write.table]{utils::write.csv()}} function.}

\item{path:}{a path to the folder to save the phyloseq object}
}
\value{
One to four csv tables (refseq.csv, otu_table.csv, tax_table.csv, sam_data.csv)
and if present a phy_tree in Newick format
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
\dontrun{
write_phyloseq(data_fungi, path = "phyloseq")
write_phyloseq(data_fungi, path = "phyloseq", one_file_ASV = TRUE)
}

}
